### 2025-02-13 #685 @kmartinet
SSARP: Easily Create Species- and Speciation- Area Relationships Using Web Databases
Submitting Author Name: Kristen Martinet
Submitting Author Github Handle: @kmartinet
Repository:  https://github.com/kmartinet/SSARP
Version submitted: 0.2.0
Submission type: Standard

Version accepted: TBD

---

```
Package: SSARP
Title: SSARP (Species-/Speciation-Area Relationship Projector)
Description: Create Species- and Speciation-Area Relationships for any taxa.
Authors@R: person("Kristen", "Martinet", email = "kmartinet@outlook.com", role = c("aut", "cre"))
Maintainer: Kristen Martinet <kmartinet@outlook.com>
Version: 0.2.0
  ape,
  checkmate,
  cli,
  Dict,
  epm,
  httr,
  mapdata,
  maps,
  rgbif,
  segmented,
  stringi,
  tidyr
License: GPL (>=2)
    knitr (>= 1.46)
```

The _SSARP_ R package provides a smooth workflow for obtaining occurrence records for taxa of interest from GBIF (through the use of [rgbif](https://github.com/ropensci/rgbif) utilities), associating GPS points with appropriate landmasses, getting the area of those landmasses, calculating speciation rates, and creating species- and speciation-area relationships.

_SSARP_'s target audience includes researchers who want to create species- and speciation-area relationships in order to visualize trends in biodiversity across island (or island-like) systems globally. Species- and speciation-area relationships are useful for determining whether biodiversity is distributed in an unusual way in a given island system (e.g., species richness decreasing as islands get larger instead of increasing). An unusual trend in these relationships may be indicative of habitat loss or a negative impact of non-native species.

No other R packages provide the workflow that _SSARP_ provides, but there are other R packages that provide some of the functionality. The [sars](https://cran.r-project.org/web/packages/sars/vignettes/sars-r-package.html) package fits several models for species-area relationships. The [CoordinateCleaner](https://github.com/ropensci/CoordinateCleaner) package provides a robust suite of methods to clean occurrence record data (e.g., filtering out invalid GPS locations).
