### 2025-02-15 #685 @maurolepore
Dear @kmartinet thanks for your submission. I'll discuss it with the editorial board. I'll be back in a few days.

For the record I leave a first quick preliminary check so that I can inform the other editors what I see. If we take it, the handling editor may do more complete checks.

### Quick preliminary checks:

      > The case is not made explicitely, but it's implicit form it's goal and the multiple
      > processes it automates. It would be nice to see the case for the package
      > at the very top of README, following the goal.
      > The functions are not grouped.
      > Function names could be more consistent in letter-case and verb-noun structure.
      > README shows no example.
      > The vignettes seem meaningful but show no code output.
      > The few helpfiles I explored randomly show no code output.
    > I see some spatial dependencies that may be hard to install. On my laptop:
    ```
    ✖ Failed to build terra 1.8-21 (1.6s)
    ```
    > From README I expect to interact with an API but in my quick check I haven't seen HTTP tests or dependencies.
    > I don't see instructions to connect to extarnal resources. No token needed?

  ---

* R/f.R alligns with tests/testthat/test-f. :-)
* I see some style issues. Consider https://style.tidyverse.org/ and https://styler.r-lib.org/
* The code I've seen shows no output. Does it all run?
* At least some calls to `expect_error()` are inspecific (pick any error).
* I see a few tests for messages. Are they easy to suppress (e.g. in tests)?
