### 2025-02-18 #685 @maurolepore
@kmartinet you're very welcome. I already advertised the package among editors and I'm still waiting for someone to become available.

If you're available and want to make some progress, then you may start by addressing the issues that take the least effort and make the biggest difference in helping a reviewer understand the package as quickly as possible.

The most obvious place is README. Adding a reproducible example and showing output can help a reviewer understand what the package actually does even before they try to install it.

Another good investment of your time might be on installation. Spatial dependencies can be tricky to install. If you help the reviewers hit the ground running, then the limited time they have can be spent in the kind of  feedback that would benefit the package the most.
