+++ README.md
@@ -5,7 +5,9 @@
   [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
   <!-- badges: end -->
 
-SSARP (Species-/Speciation-Area Relationship Projector) is an R package that provides a suite of functions to help users **create speciation- and species-area relationships for island-dwelling taxa** using occurrence data from GBIF (Global Biodiversity Information Facility) accessed via GBIF's API or the user's own occurrence data.
+SSARP (Species-/Speciation-Area Relationship Projector) is an R package that provides a suite of functions to help users **create speciation- and species-area relationships for island-dwelling taxa** using occurrence data from GBIF (Global Biodiversity Information Facility) accessed via GBIF's API or the user's own occurrence data. 
+
+Species- and speciation-area relationships are useful for determining whether biodiversity is distributed in an unusual way in a given island system (e.g., species richness decreasing as islands get larger instead of increasing). An unusual trend in these relationships may be indicative of habitat loss or a negative impact of non-native species.
 
 Please find the [bioRxiv preprint of the manuscript associated with *SSARP* here!](https://www.biorxiv.org/content/10.1101/2024.12.31.630948v1)
 
@@ -18,7 +20,87 @@ library(devtools)
 install_github("kmartinet/SSARP")
 ```
 
-### Basic Workflow - Using data from GBIF to create a species-area relationship plot
+### Example: Creating a Species-Area Relationship
+A species-area relationship (SAR) visualizes the relationship between species richness (the number of species) and the area of the land mass on which the species live. This brief example covers the *SSARP* workflow for creating a SAR, and more detailed explanations of the code and methods can be found [in the Articles on the SSARP pkgdown website](https://kmartinet.github.io/SSARP/index.html).
+
+In order to construct a species-area relationship with *SSARP*, we will:
+
+- Gather occurrence data from GBIF
+- Filter out invalid occurrence records
+- Find areas of pertinent land masses
+- Create a species-area relationship
+
+#### Step 1: Gather occurrence data from GBIF
+In this step, we will find the unique identifying key associated with a taxon of interest (the lizard genus *Anolis* in this case) and use that key to access occurrence points for that taxon from GBIF. We will restrict the returned occurrence records to the Caribbean islands using the "geometry" parameter in the "getData" function.
+
+```
+library(SSARP)
+
+# Get the GBIF key for the Anolis genus
+key <- getKey(query = "Anolis", rank = "genus")
+
+# Get data for Anolis from GBIF in a specified polygon around Caribbean islands
+dat <- getData(key = key, limit = 10000,
+  geometry = 'POLYGON((-84.8 23.9, -84.7 16.4, -65.2 13.9, -63.1 11.0, -56.9 15.5, -60.5 21.9, -79.3 27.8, -79.8 24.8, -84.8 23.9))')
+```
+
+#### Step 2: Filter out invalid occurrence records
+Now that we have a dataframe that includes the first 10,000 records for *Anolis* within a specified polygon around the Caribbean islands, we will filter that data to include only occurrence records that are on land. Some occurrence records might have GPS points that are in the ocean instead of on an island, so it is important to exclude these invalid records.
+
+```
+# Find land mass names and exclude records not on land
+land_dat <- findLand(occurrences = dat)
+```
+
+#### Step 3: Find areas of pertinent land masses
+Next, we will use the dataframe of occurrence records with their associated land mass names (the "land_dat" object created above) with the "findAreas" function to find the area of each land mass.
+```
+# Use the land mass names to get their areas
[…]
