### 2025-02-18 aecaa12
Made help file examples more robust, fixed quickSARP tests
