+++ R/speciationMS.R
@@ -7,7 +7,15 @@
 #' @return A dataframe that includes speciation rates for each island in the user-provided occurrence record dataframe
 #' @examples 
 #' \dontrun{
-#' occ_speciation <- speciationMS(tree, "epithet", occs)
+#' key <- getKey(query = "Anolis", rank = "genus")
+#' dat <- getData(key = key, limit = 100)
+#' land <- findLand(occurrences = dat)
+#' areas <- findAreas(occs = land)
+#' 
+#' # Assuming that the user has a tree file called "anolis.tree"
+#' tree <- ape::read.tree("anolis.tree")
+#' 
+#' occ_speciation <- speciationMS(tree = tree, label_type = "epithet", occurrences = areas)
 #' }
 #' @importFrom dplyr mutate case_when
 #' @importFrom ape subtrees node.depth.edgelength
