+++ R/getKey.R
@@ -6,10 +6,10 @@
 #' @return The taxon key that GBIF needs to search its database for occurrence records
 #' @examples 
 #' \dontrun{
-#' key <- getKey("Phelsuma", "genus")
-#' key <- getKey("Tamias ruficaudus ruficaudus", "subspecies")
-#' key <- getKey("Naesiotus akamatus", "species")
-#' key <- getKey("Araneae", "order")
+#' key <- getKey(query = "Phelsuma", rank = "genus")
+#' key <- getKey(query = "Tamias ruficaudus ruficaudus", rank = "subspecies")
+#' key <- getKey(query = "Naesiotus akamatus", rank = "species")
+#' key <- getKey(query = "Araneae", rank = "order")
 #' }
 #' @importFrom rgbif name_suggest
 #' @importFrom checkmate assertString

[…]
