### 2025-02-19 af3c1af
Removed epm dependency, printed dataframes in SAR vignette
