+++ vignettes/Create_SAR.Rmd
@@ -32,6 +32,13 @@ library(SSARP)
 
 # Get the GBIF key for the Anolis genus
 key <- getKey(query = "Anolis", rank = "genus")
+
+# Print key
+key
+```
+
+```
+## [1] 8782549
 ```
 
 Now that we have the key associated with *Anolis*, we will use this key with *SSARP*'s "getData" function to gather georeferenced occurrence records from GBIF. The “limit” parameter will be set to 10000 in this case for a quick illustration of *SSARP*’s functionality, but this parameter can be as large as 100,000 (the hard limit from *rgbif* for the number of records returned).
@@ -45,6 +52,27 @@ We are only interested in occurrence records for island-dwelling anole lizards l
 ``` r
 # Get data for Anolis from GBIF in the specified polygon
 dat <- SSARP::getData(key = key, limit = 10000, geometry = 'POLYGON((-84.8 23.9, -84.7 16.4, -65.2 13.9, -63.1 11.0, -56.9 15.5, -60.5 21.9, -79.3 27.8, -79.8 24.8, -84.8 23.9))')
+
+# Print first 5 lines of dat
+head(dat, n = 5)
+```
+
+```
+## # A tibble: 5 × 160
+##   key    scientificName decimalLatitude decimalLongitude issues datasetKey publishingOrgKey installationKey
+##   <chr>  <chr>                    <dbl>            <dbl> <chr>  <chr>      <chr>            <chr>          
+## 1 50070… Anolis distic…            19.0            -69.0 cdc,c… 50c9509d-… 28eb1a3f-1c15-4… 997448a8-f762-…
+## 2 50070… Anolis everma…            18.3            -66.3 cdc,c… 50c9509d-… 28eb1a3f-1c15-4… 997448a8-f762-…
+## 3 50075… Anolis crista…            18.4            -66.0 cdc,c… 50c9509d-… 28eb1a3f-1c15-4… 997448a8-f762-…
+## 4 50081… Anolis sagrei…            18.4            -64.5 cdc,c… 50c9509d-… 28eb1a3f-1c15-4… 997448a8-f762-…
+## 5 50370… Anolis leachi…            17.1            -61.8 cdc,c… 50c9509d-… 28eb1a3f-1c15-4… 997448a8-f762-…
+## # ℹ 152 more variables: hostingOrganizationKey <chr>, publishingCountry <chr>, protocol <chr>,
+## #   lastCrawled <chr>, lastParsed <chr>, crawlId <int>, basisOfRecord <chr>, occurrenceStatus <chr>,
+## #   taxonKey <int>, kingdomKey <int>, phylumKey <int>, classKey <int>, familyKey <int>, genusKey <int>,
+## #   speciesKey <int>, acceptedTaxonKey <int>, acceptedScientificName <chr>, kingdom <chr>, phylum <chr>,
+## #   family <chr>, genus <chr>, species <chr>, genericName <chr>, specificEpithet <chr>, taxonRank <chr>,
+## #   taxonomicStatus <chr>, iucnRedListCategory <chr>, dateIdentified <chr>,
+## #   coordinateUncertaintyInMeters <dbl>, continent <chr>, stateProvince <chr>, year <int>, month <int>, …
 ```
 
 ## Finding Land Mass Names and Areas
@@ -55,9 +83,54 @@ Once the occurrence data is returned, we will use each occurrence record’s GPS
 # Find land mass names
 land_dat <- findLand(occurrences = dat)
 
+# Print first 5 lines of land_dat
+head(land_dat, n = 5)
+```
+
+```
+##                                  SpeciesName  Genus      Species  Longitude  Latitude       First Second
[…]
