+++ NAMESPACE
@@ -30,7 +30,6 @@ importFrom(cli,cli_alert_warning)
 importFrom(dplyr,case_when)
 importFrom(dplyr,count)
 importFrom(dplyr,mutate)
-importFrom(epm,DRstat)
 importFrom(graphics,abline)
 importFrom(graphics,points)
 importFrom(httr,GET)
