+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   Create_SAR: Create_SAR.html
   Create_SpAR: Create_SpAR.html
-last_built: 2025-02-19T20:07Z
+last_built: 2025-02-19T20:22Z
 urls:
   reference: https://kmartinet.github.io/SSARP/reference
   article: https://kmartinet.github.io/SSARP/articles
