### 2025-02-19 #685 @kmartinet
Hi @maurolepore , thanks so much for the guidance! I have more explicitly written the case for the package at the top of the README and added a reproducible example for creating a species-area relationship (SAR). Additionally, the SAR vignette now includes more in-depth discussion about the structure of the objects created by each step of the workflow.

I also re-worked the "speciationDR" function to remove _SSARP_'s dependency on the _epm_ package, which was the only reason why _terra_ was originally part of _SSARP_'s installation.
