### 2025-02-25 d648dd4
Streamlining how speciationMS interacts with SpeARP
