+++ tests/testthat/test-SpeARP.R
@@ -7,8 +7,8 @@ occ_zero_bp <- as.data.frame(cbind(areas_0, rate_0))
 colnames(occ_zero_bp) <- c("areas", "rate")
 
 # One breakpoint
-areas <- c(9710687500,301187500,83104562500,1.19e+09,643812500,228562500,187500,1.22e+11,263875000,1196500000,14250000,321875000,268312500,12225750000,812937500,1.26e+08,528125000,16937500,33625000,3125000,103625000,180500000,22812500,329750000,121812500,212375000,6117437500,28125000,2.2e+07,309375000,433687500,240937500,44875000,1765125000,1.42e+08,112375000,44312500,56062500)
-rate <- c(0.0568884629420592,0,0.0367002192918066,0,0,0,0,0,0,0,0,0,0,0.0448931329321796,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0)
+areas <- c(9710687500,301187500,12225750000,83104562500,1.19e+09,643812500,228562500,187500,1.22e+11,263875000,329750000,1196500000,14250000,1.26e+08,321875000,268312500,812937500,528125000,16937500,33625000,3125000,103625000,180500000,22812500,121812500,212375000,6117437500,28125000,2.2e+07,309375000,433687500,240937500,44875000,1765125000,1.42e+08,112375000,44312500,56062500)
+rate <- c(1.0585377376259,1,1.04591607995137,1.03738198711277,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1)
 occ_one_bp <- as.data.frame(cbind(areas, rate))
 
 # Two breakpoints
@@ -20,6 +20,7 @@ rate_2 <- c(0, 0, 0, 0,
              0, 1, 2, 3, 4,
              0, 1, 2, 3, 4,
              0, 1, 2, 3, 4)
+rate_2 <- rate_2 + 1
 areas_2 <- c(100, 200, 300, 400,
            500, 500,
            600, 600, 600,
@@ -46,9 +47,13 @@ test_that("Inputting a non-numeric into npsi throws an error", {
 })
 
 test_that("The SpeARP function returns a list (zero breakpoints)", {
-  expect_type(SpeARP(occ_zero_bp, npsi = 0, MS = FALSE), "list")
+  expect_type(SpeARP(occ_zero_bp, npsi = 0), "list")
 })
 
 test_that("The SpeARP function returns a list (one breakpoint)", {
-  expect_type(SpeARP(occ_one_bp, npsi = 1, MS = TRUE), "list")
+  expect_type(SpeARP(occ_one_bp, npsi = 1), "list")
+})
+
+test_that("The SpeARP function returns a list (two breakpoints)", {
+  expect_type(SpeARP(occ_two_bp, npsi = 2), "list")
 })
