### 2025-02-26 bad537f
Add findShapeAreas function to use custom SpatVectors
