+++ NAMESPACE
@@ -6,6 +6,7 @@ export(SARP)
 export(SpeARP)
 export(findAreas)
 export(findLand)
+export(findShapeAreas)
 export(getData)
 export(getKey)
 export(getSources)
@@ -21,6 +22,7 @@ import(maps)
 importFrom(ape,node.depth.edgelength)
 importFrom(ape,read.tree)
 importFrom(ape,subtrees)
+importFrom(checkmate,assertClass)
 importFrom(checkmate,assertDataFrame)
 importFrom(checkmate,assertLogical)
 importFrom(checkmate,assertNumeric)
@@ -39,8 +41,13 @@ importFrom(rgbif,name_suggest)
 importFrom(rgbif,occ_search)
 importFrom(segmented,seg.control)
 importFrom(segmented,segmented)
+importFrom(sf,st_area)
+importFrom(sf,st_as_sf)
 importFrom(stats,AIC)
 importFrom(stats,aggregate)
 importFrom(stats,lm)
 importFrom(stringi,stri_unescape_unicode)
+importFrom(terra,extract)
+importFrom(terra,subset)
+importFrom(terra,vect)
 importFrom(tidyr,separate)
