+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: SSARP (Species-/Speciation-Area Relationship Projector)
 Description: Create Species- and Speciation-Area Relationships for any taxa.
 Authors@R: person("Kristen", "Martinet", email = "kmartinet@outlook.com", role = c("aut", "cre"))
 Maintainer: Kristen Martinet <kmartinet@outlook.com>
-Version: 0.2.0
+Version: 0.3.0
 URL: https://github.com/kmartinet/SSARP, https://kmartinet.github.io/SSARP/
 BugReports: https://github.com/kmartinet/SSARP/issues
 Imports: 
@@ -18,7 +18,9 @@ Imports:
   maps,
   rgbif,
   segmented,
+  sf,
   stringi,
+  terra,
   tidyr
 LazyData: true
 Encoding: UTF-8
