### 2025-03-03 a0d9a9b
Added test cases to findShapeAreas
