### 2025-03-14 2709f18
Added islands, also edited findShapeAreas example
