### 2025-03-14 1e7fd42
SpeARP returns AIC scores just like SARP now
