+++ R/SpeARP.R
@@ -58,6 +58,8 @@ SpeARP <- function(occurrences, npsi = 1) {
   aic_scores <- list()
   # Linear model is already created above
   aic_scores[1] <- AIC(linear)
+  # Name the list element in case the user wants to examine the AIC scores
+  names(aic_scores)[1] <- "Linear"
   
   # Create a segmented model for each level of npsi specified by the user
   # This only makes sense if the user does not specify that they want zero breakpoints
@@ -65,6 +67,8 @@ SpeARP <- function(occurrences, npsi = 1) {
     for(i in seq(npsi)){
       seg <- segmented(linear, seg.Z = ~x, npsi = i, control = seg.control(display = FALSE))
       aic_scores[i+1] <- AIC(seg)
+      # Name the list element
+      names(aic_scores)[i+1] <- paste0("BP", i)
     }
   }
   
@@ -84,7 +88,7 @@ SpeARP <- function(occurrences, npsi = 1) {
     
     summary_line <- summary(linear)
     
-    result <- list("summary" = summary_line, "linObj" = linear, "aggDF" = dat)
+    result <- list("summary" = summary_line, "linObj" = linear, "aggDF" = dat, "AICscores" = aic_scores)
     
     class(result) <- "SAR"
     
@@ -108,7 +112,7 @@ SpeARP <- function(occurrences, npsi = 1) {
     # Save the summary as an object to add to the result list
     summary_seg <- summary(seg)
     
-    result <- list("summary" = summary_seg, "segObj" = seg, "aggDF" = dat)
+    result <- list("summary" = summary_seg, "segObj" = seg, "aggDF" = dat, "AICscores" = aic_scores)
     
     class(result) <- "SAR"
     
@@ -126,7 +130,7 @@ SpeARP <- function(occurrences, npsi = 1) {
     # Save the summary as an object to add to the result list
     summary_seg <- summary(seg)
     
-    result <- list("summary" = summary_seg, "segObj" = seg, "aggDF" = dat)
+    result <- list("summary" = summary_seg, "segObj" = seg, "aggDF" = dat, "AICscores" = aic_scores)
     
     class(result) <- "SAR"
     
