### 2025-03-17 3f150a0
Added getPresenceAbsence function for presence/absence matrix
