+++ man/getPresenceAbsence.Rd
@@ -1,0 +1,33 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/getPresenceAbsence.R
+\name{getPresenceAbsence}
+\alias{getPresenceAbsence}
+\title{Create a presence-absence dataframe for a given occurrence record dataframe}
+\usage{
+getPresenceAbsence(occs)
+}
+\arguments{
+\item{occs}{The dataframe output by SSARP::findAreas, or if using a custom dataframe, ensure that it has the following named columns:
+\itemize{
+\item "areas" containing the areas associated with the land masses of interest
+\item "Species" containing the names of the species living on those islands
+\item A 6th column containing locality information. In the SSARP workflow, this column is called "First" and contains the country name
+\item A 7th column containing locality information. In the SSARP workflow, this column is called "Second" and contains a province or island name
+\item An 8th column containing locality information. In the SSARP workflow, this column is called "Third" and contains the island name if the 7th column does not contain the island name
+}}
+}
+\value{
+A dataframe with a row for each island in the given occurrence record dataframe and a column for each species. Within each species column, a 1 represents the presence of that species on the island corresponding to the given row, and a 0 represents the absence of that species on the island corresponding to the given row.
+}
+\description{
+Use a dataframe output by SSARP::findAreas to determine which species occur on specific islands by creating a presence-absence dataframe. A 1 represents presence and a 0 represents absence.
+}
+\examples{
+\dontrun{
+key <- getKey(query = "Anolis", rank = "genus")
+dat <- getData(key = key, limit = 100)
+land <- findLand(occurrences = dat)
+areas <- findAreas(occs = land)
+pres_abs <- getPresenceAbsence(areas)
+}
+}
