+++ NAMESPACE
@@ -9,6 +9,7 @@ export(findLand)
 export(findShapeAreas)
 export(getData)
 export(getKey)
+export(getPresenceAbsence)
 export(getSources)
 export(get_island_areas)
 export(quickSARP)
@@ -37,6 +38,7 @@ importFrom(graphics,points)
 importFrom(httr,GET)
 importFrom(httr,content)
 importFrom(httr,user_agent)
+importFrom(reshape2,dcast)
 importFrom(rgbif,name_suggest)
 importFrom(rgbif,occ_search)
 importFrom(segmented,seg.control)
