### 2025-04-07 #685 @maurolepore
@kmartinet thanks for the reminder. I re-announced the package to our editors and I hope someone responds soon.

Sorry for being vague about the status. Other editors agree that the package is in scope and also that most of the minor issues I noted should be addressed before engaging reviewers. So great you've worked on them.

RE showing output, I meant that it would be useful to see not just the code but also the output that the code produces.

Eg the code '1 + 1' outputs '2'. Last time I checked I saw '1 + 1' but not '2'.

In other words, the code in README.Rmd should be in executable code chunks, and and README.Rmd should be rendered with  'eval = TRUE'.

Does that make sense?
