+++ README.md
@@ -1,27 +1,54 @@
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
 # SSARP
-  <!-- badges: start -->
-  [![Codecov test coverage](https://codecov.io/gh/kmartinet/SSARP/branch/main/graph/badge.svg)](https://codecov.io/gh/kmartinet/SSARP?branch=main)
-  [![R-CMD-check](https://github.com/kmartinet/SSARP/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/kmartinet/SSARP/actions/workflows/R-CMD-check.yaml)
-  [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-  <!-- badges: end -->
 
-SSARP (Species-/Speciation-Area Relationship Projector) is an R package that provides a suite of functions to help users **create speciation- and species-area relationships for island-dwelling taxa** using occurrence data from GBIF (Global Biodiversity Information Facility) accessed via GBIF's API or the user's own occurrence data. 
+<!-- badges: start -->
 
-Species- and speciation-area relationships are useful for determining whether biodiversity is distributed in an unusual way in a given island system (e.g., species richness decreasing as islands get larger instead of increasing). An unusual trend in these relationships may be indicative of habitat loss or a negative impact of non-native species.
+[![Codecov test
+coverage](https://codecov.io/gh/kmartinet/SSARP/branch/main/graph/badge.svg)](https://codecov.io/gh/kmartinet/SSARP?branch=main)
+[![R-CMD-check](https://github.com/kmartinet/SSARP/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/kmartinet/SSARP/actions/workflows/R-CMD-check.yaml)
+[![Project Status: Active – The project has reached a stable, usable
+state and is being actively
+developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+<!-- badges: end -->
 
-Please find the [bioRxiv preprint of the manuscript associated with *SSARP* here!](https://www.biorxiv.org/content/10.1101/2024.12.31.630948v1)
+SSARP (Species-/Speciation-Area Relationship Projector) is an R package
+that provides a suite of functions to help users **create speciation-
+and species-area relationships for island-dwelling taxa** using
+occurrence data from GBIF (Global Biodiversity Information Facility)
+accessed via GBIF’s API or the user’s own occurrence data.
 
-### Installation
+Species- and speciation-area relationships are useful for determining
+whether biodiversity is distributed in an unusual way in a given island
+system (e.g., species richness decreasing as islands get larger instead
+of increasing). An unusual trend in these relationships may be
+indicative of habitat loss or a negative impact of non-native species.
 
-To install *SSARP*, use the "install_github" function from the *devtools* package:
-```
+Please find the [bioRxiv preprint of the manuscript associated with
+*SSARP*
+here!](https://www.biorxiv.org/content/10.1101/2024.12.31.630948v1)
+
+## Installation
+
+To install *SSARP*, use the “install_github” function from the
+*devtools* package:
+
+``` r
+# install.packages("devtools")
 library(devtools)
 
 install_github("kmartinet/SSARP")
 ```
 
-### Example: Creating a Species-Area Relationship
[…]
