+++ README.Rmd
@@ -1,0 +1,145 @@
+---
+output: github_document
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
+)
+```
+
+# SSARP
+
+<!-- badges: start -->
+  [![Codecov test coverage](https://codecov.io/gh/kmartinet/SSARP/branch/main/graph/badge.svg)](https://codecov.io/gh/kmartinet/SSARP?branch=main)
+  [![R-CMD-check](https://github.com/kmartinet/SSARP/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/kmartinet/SSARP/actions/workflows/R-CMD-check.yaml)
+  [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+<!-- badges: end -->
+
+SSARP (Species-/Speciation-Area Relationship Projector) is an R package that provides a suite of functions to help users **create speciation- and species-area relationships for island-dwelling taxa** using occurrence data from GBIF (Global Biodiversity Information Facility) accessed via GBIF's API or the user's own occurrence data. 
+
+Species- and speciation-area relationships are useful for determining whether biodiversity is distributed in an unusual way in a given island system (e.g., species richness decreasing as islands get larger instead of increasing). An unusual trend in these relationships may be indicative of habitat loss or a negative impact of non-native species.
+
+Please find the [bioRxiv preprint of the manuscript associated with *SSARP* here!](https://www.biorxiv.org/content/10.1101/2024.12.31.630948v1)
+
+## Installation
+
+To install *SSARP*, use the "install_github" function from the *devtools* package:
+
+``` r
+# install.packages("devtools")
+library(devtools)
+
+install_github("kmartinet/SSARP")
+```
+
+## Example: Creating a Species-Area Relationship
+
+A species-area relationship (SAR) visualizes the relationship between species richness (the number of species) and the area of the land mass on which the species live. This brief example covers the *SSARP* workflow for creating a SAR, and more detailed explanations of the code and methods can be found [in the Articles on the SSARP pkgdown website](https://kmartinet.github.io/SSARP/index.html).
+
+In order to construct a species-area relationship with *SSARP*, we will:
+
+- Gather occurrence data from GBIF
+- Filter out invalid occurrence records
+- Find areas of pertinent land masses
+- Create a species-area relationship
+
+#### Step 1: Gather occurrence data from GBIF
+In this step, we will find the unique identifying key associated with a taxon of interest (the lizard genus *Anolis* in this case) and use that key to access occurrence points for that taxon from GBIF. We will restrict the returned occurrence records to the Caribbean islands using the "geometry" parameter in the "getData" function.
+
+```{r, warning = FALSE, message = FALSE}
+library(SSARP)
+
+# Get the GBIF key for the Anolis genus
+key <- getKey(query = "Anolis", rank = "genus")
+
+# Print the key
+key
+
+```
+The GBIF key for *Anolis* is 8782549. We will use this key, along with a Well-Known Text (WKT) polygon around Caribbean islands, to get data for *Anolis* from GBIF using the "getData" function from *SSARP*.
+
+```{r, warning = FALSE, message = FALSE}
+
+# Get data for Anolis from GBIF in a specified polygon around Caribbean islands
[…]
