### 2025-04-07 #685 @kmartinet
Thanks so much, @maurolepore !

I've edited my README workflow to match your specifications. The example in the README now shows the first 5 lines of each dataframe that is created in each step of the _SSARP_ workflow, along with a brief description for each output.
