### 2025-04-09 #685 @maurolepore
Dear @kmartinet thanks again for sharing your work with rOpenSci and for responding to my feedback.

@jhollist kindly offered to be the handling editor. The bot command failed but he's aware and should be here sometime soon.
