### 2025-04-25 #685 @jhollist
@kmartinet I see that you and @maurolepore had some early back and forth and looks like you have addressed most of his initial suggestions.  I re-ran the checks and nothing major jumps out, but a few notes.

- Don't worry about the NAs in the Package Dependency checks.  I did a quick look and looks like you have these called out appropriately with @importFrom.
- Take a look at the `lintr` results from in 3b at https://github.com/ropensci/software-review/issues/685#issuecomment-2828125296 and address as many as you can.  These can be taken care of while I invite reviewers.

Other than these small things, we should be good to move forward.  I'll start looking for reviewers today.
