### 2025-04-26 7c50713
Addressed lintr character flags and 1:length(...) flags
