### 2025-04-26 #685 @kmartinet
Thank you so much, @jhollist !

I've addressed all of the lintr suggestions to edit lines that have greater than 80 characters, with a few exceptions:
-  getKey.R line 39: I'd like to preserve the look of the alert text in this line.
- getSources.R lines 15 and 16: Links in the bulleted list in the documentation reference section make these lines long, and I'd like to keep these helpful links here.
- test-speciationBAMM.R, test-speciationDR, and test-speciationMS line 2: A phylogenetic tree is read from a link here, and the link pushes this line over 80 characters. I'd like to preserve this method for reading in this tree.

lintr also suggested replacing the use of 1:length(...) in two places. I changed the logic for these to use seq_len instead.
