### 2025-04-29 79d6d18
Added NEWS.md and rOpenSci review status badge
