+++ NEWS.md
@@ -1,0 +1,10 @@
+SSARP 0.2.0 (2025-04-29)
+=========================
+
+### NEW FEATURES
+
+  * Added NEWS file
+  
+### DOCUMENTATION FIXES
+  * Added badge for status at rOpenSci software peer review to README
+  
