### 2025-05-04 #685 @maurolepore
Dear all I'm wrapping my EiC rotation and handing it over to the next EiC along with some short notes.

I see a lot of great activity. Thanks so much for your work!
