### 2025-05-19 9030bf7
Removed references to MS parameter in SpAR vignette
