### 2025-06-12 #685 @txm676
Yes, with the caveat that most examples are set to not be run

With the caveat that it appears some are not testing the output is
actually correct.

---

The package has been well put together, and I didn't come across any major issues with installation or use. I work a lot on SAR analyses (and write the sars R package), so can say with confidence (or at least misplaced confidence) that the package is fulfilling a useful role and will make a nice contribution to the island biogeographer's tool kit. The accompanying paper is very useful. I have some comments below you may wish to consider.

#### R checks

When running on my PC, there was one note:

Running R checks resulted in one note:
checking files in 'vignettes' ... NOTE
  The following directory looks like a leftover from 'knitr':
    'figure'
  Please remove from your package.

#### Model fitting and output

For the two main model fitting functions, would it be better to return all fitted models as well as the best (in case the user wants them for something)?

Do the model fitting functions work if the user provide islands with zero species (i.e., not data from GBIF), given the log-transformation?

Unless I missed it, there isn’t a way to generate / return a sort of standard island-species richness dataframe, which would be very useful in downstream analyses (e.g., if you wanted to fit other models than segmented relationships). Presumably this is created internally within `SARP`, so could it also be returned in the output of this function (and not just the log-transformed version)?

Code format / structure mostly looks good, but there are some extra spaces and blank lines in places.

There isn't a fully consistent function naming notation used, e.g., `SARP` doesn't follow the same notation as many of the others. Also should the package name be in lower case?

For the `SARP` function, there should be an option to not return the plot.

Not essential, but do you need two main functions for building the SAR (i.e, `SARP` and `SpeARP`) if they are fitting the same models? The input is different, but you could just deal with that internally and have one function (to avoid code repetition). Or if you want to keep them distinct, can the segmented and plotting code be part of an internal function that is then called by both to avoid the repetition.

Do you need the `getData` functions, or could the users just use `rgbif::occ_download()` directly? I partly ask this because there are many other arguments in the original functions that can help filter searches that aren’t accessible through your functions. At the very least I think it would be helpful to allow users to provide those arguments (e.g., using `…`)

Similarly with the plot functions, it would be useful to allow authors to provide more arguments for editing these plots (if possible), or direct users to the original plot function documentation (and provide output that can be fed into it directly)

Just a little bit more info about each of the three speciation rate metrics in the vignette would be helpful. Obviously readers would need to visit the original papers but I think some more information on each would still be useful. Or at minimum direct them to Table 2 in your paper and the methods.

The two SAR building functions could do with some more information in their docs, particularly `SpeARP` (where all three speciation rate functions should be listed, ideally with links through to their associated docs). A lot of useful information is in the paper, so perhaps this could just be transferred across.

An example involving the user providing their own dataset (rather than Gbif) would be a useful addition to the vignette (unless I missed this!)

A common approach to island SAR analysis to build the SAR for islands within an archipelago. I assume for this to work here, you would need to provide a geometry argument that delineates the archipelago; perhaps this could be made clearer (given how common it is as an approach).

#### Input data

Can you provide `anolis.tree` (Patton_Anolis_Trimmed) as a data object (or just a simulated example tree) or place it in inst/extdata? As currently it is a bit long-winded to load it in. This would also then enable the relevant unit test to actually work with the standard input rather than manually building a tree as part of the test.

I am a bit worried about the max threshold of 100,000 for GBIF searches, not so much that it exists, but more than people may just use your functions without realising this. This is made clearer in the paper than in the package documentation, so perhaps this caveat could be made more explicit. Also in the paper is a discussion on potential ways round this (e.g., searching by species) which could also perhaps be clearer in the package documentation. Another option is to search by individual island and iterate across islands, rather than as a whole set of polygons (which I assume is what is being done here?).

#### Unit tests

This is potentially a wider issue with the unit tests, as several are not really testing the functions in the way that they would be used. Also I think it would be a good idea to add some tests for actual output values (rather than just successful competition / erroring) to make sure future changes to the code don't affect the actual output of the functions.

#### Other points

The statement at the top of the README: "Species- and speciation-area relationships are useful for determining whether biodiversity is distributed in an unusual way in a given island system". While true, I think there are more standard uses of SARs that would be better to highlight here (e.g., testing various biodiversity theories), which you do well in the paper.

This is completely optional obviously (as it is my own package), but I thought it could be nice to direct users the sars R package in the README for functions for fitting other models, i.e., to build the SAR data using SSARP and then with the potential to feed through into sars. I notice you do it in the paper already (thanks!). This is also why the previous comment is important, as having the data in that format will aid in this (and not just with sars, with any subsequent package). I will also add a link for SSARP in the sars README as there are no functions in sars for building SAR data. I think the packages nicely complement themselves in this way and it could be useful to highlight this (an often repeated criticism with R is there are so many packages for doing the same type of thing, with little understanding of how they link together). But as I say, just an idea, not essential!

A comment related to the paper: Not sure about this statement, as there have been a lot of global ISAR comparison studies: Global comparisons of island systems are not common in the island biogeography literature because of the difficulty associated with gathering and processing the large amount of data needed to make informed conclusions
