### 2025-07-05 49cda1c
Documentation improvements and testthat fixes
