+++ man/find_land.Rd
@@ -7,16 +7,16 @@
 find_land(occurrences, fillgaps = FALSE)
 }
 \arguments{
-\item{occurrences}{The dataframe output by getData (or if using a custom
-dataframe, ensure that it has the following columns: decimalLongitude,
-decimalLatitude, acceptedScientificName, genericName, specificEpithet,
-datasetKey). The "datasetKey" column is important for GBIF records and
-identifies the dataset to which the occurrence record belongs. Custom
-dataframes without this style of data organization should fill the column
-with placeholder values.}
+\item{occurrences}{The dataframe output by \code{SSARP::get_data()} (or if using
+a custom dataframe, ensure that it has the following columns:
+decimalLongitude, decimalLatitude, acceptedScientificName, genericName,
+specificEpithet, datasetKey). The "datasetKey" column is important for GBIF
+records and identifies the dataset to which the occurrence record belongs.
+Custom dataframes without this style of data organization should fill the
+column with placeholder values.}
 
 \item{fillgaps}{(logical) Attempt to use Photon API to fill in gaps left by
-mapdata::map.where (TRUE) or only mapdata::map.where results
+\code{mapdata::map.where()} (TRUE) or only \code{mapdata::map.where()} results
 (FALSE, default). While it is powerful, the Photon API does not have a
 standard location for island names in its returned information, so using it
 will likely require the returned dataframe to be cleaned by the user.}
