### 2025-07-05 8bc92c5
Instead of testthat::test_path, attempting system.file again
