+++ tests/testthat/test-estimate_DR.R
@@ -3,7 +3,9 @@
 #                               "Patton_Anolis_trimmed.tree", 
 #                               package = "SSARP"))
 tree <- 
-  ape::read.tree(testthat::test_path("testdata/Patton_Anolis_trimmed.tree"))
+  ape::read.tree(system.file("testthat/testdata/Patton_Anolis_trimmed.tree", 
+                              package = "SSARP"))
+
 # Tree from: Patton, A.H., Harmon, L.J., del Rosario Castañeda, M., 
 #   Frank, H.K., Donihue, C.M., Herrel, A., & Losos, J.B. (2021). When adaptive 
 #   radiations collide: Different evolutionary trajectories between and within 
