### 2025-07-05 f79401e
Run tests from source to fix testthat fail
