### 2025-07-05 0984634
Removing install path from test-coverage
