+++ .github/workflows/test-coverage.yaml
@@ -27,15 +27,12 @@ jobs:
           extra-packages: any::covr, any::xml2
           needs: coverage
 
-      - name: Set env to run tests from source
-        run: echo "_R_CHECK_TESTS_FROM_SRC_=true" >> $GITHUB_ENV
-
       - name: Test coverage
         run: |
           cov <- covr::package_coverage(
+            type = "tests",  # Run tests directly from source tree
             quiet = FALSE,
-            clean = FALSE,
-            install_path = file.path(normalizePath(Sys.getenv("RUNNER_TEMP"), winslash = "/"), "package")
+            clean = FALSE
           )
           covr::to_cobertura(cov)
         shell: Rscript {0}
