### 2025-07-05 ca9725e
Using system.file() with new YAML block
