+++ tests/testthat/test-estimate_BAMM.R
@@ -2,9 +2,10 @@
 # tree <- ape::read.tree(system.file("extdata", 
 #                               "Patton_Anolis_trimmed.tree", 
 #                               package = "SSARP"))
-tree <- 
-  ape::read.tree(system.file("testthat/testdata/Patton_Anolis_trimmed.tree",
-                             package = "SSARP"))
+# tree <- 
+#   ape::read.tree(system.file("testthat/testdata/Patton_Anolis_trimmed.tree",
+#                              package = "SSARP"))
+tree <- ape::read.tree("https://raw.githubusercontent.com/kmartinet/SSARP/refs/heads/main/vignettes/Patton_Anolis_Trimmed.tree")
 # Tree from: Patton, A.H., Harmon, L.J., del Rosario Castañeda, M., 
 #   Frank, H.K., Donihue, C.M., Herrel, A., & Losos, J.B. (2021). When adaptive 
 #   radiations collide: Different evolutionary trajectories between and within 
