+++ .github/workflows/test-coverage.yaml
@@ -30,9 +30,9 @@ jobs:
       - name: Test coverage
         run: |
           cov <- covr::package_coverage(
-            type = "tests",  # Run tests directly from source tree
             quiet = FALSE,
-            clean = FALSE
+            clean = FALSE,
+            install_path = file.path(normalizePath(Sys.getenv("RUNNER_TEMP"), winslash = "/"), "package")
           )
           covr::to_cobertura(cov)
         shell: Rscript {0}
