### 2025-07-05 2922764
Removed extraneous bullet in NEWS.md
