+++ NEWS.md
@@ -6,7 +6,6 @@ SSARP 0.3.0 (2025-07-04)
   * Changed the names of all functions to "verb_object" structure
   * Two new example files were added to the package: `Patton_Anolis_Trimmed.tree` and `SSARP_Example_Dat.csv` to allow users to run examples involving a phylogenetic tree of *Anolis* and GBIF data for *Anolis*, respectively
   * Added "get_presence_absence" function, which creates a presence-absence matrix when given a dataframe output by `SSARP::find_areas()`
-  * 
 
 ### DOCUMENTATION FIXES
   * Function names are now in `pkg::function()` notation throughout the documentation
