### 2025-07-05 #685 @kmartinet
Hi @jhollist , @joelnitta , and @txm676 ! I am so sorry for the delayed response to your reviews. Thank you so much for your time and effort in reviewing my package. I truly appreciate your thorough work and thoughtful suggestions!

I have responded to some of your comments below, with the rest to be addressed by July 18th, given that you agree that is an acceptable extension. It has been an incredibly busy past few months for me, and I really appreciate your patience!

[The NEWS.md update for these changes](https://github.com/kmartinet/SSARP/blob/main/NEWS.md#ssarp-030-2025-07-04)

## Partial Response to @joelnitta
**Vignettes response:**
- ```Patton_Anolis_Trimmed.tree``` is now in ```inst/extdata/``` ([In commit 7f29be1](https://github.com/kmartinet/SSARP/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- The code for accessing it across the package has been updated to ```read.tree(system.file("extdata", "Patton_Anolis_Trimmed.tree", package = "SSARP"))``` with three important exceptions: `test-estimate_BAMM()`, `test-estimate_DR()`, and `test-estimate_MS()`. The test-coverage.yaml workflow kept failing due to it not being able to find the file (even though it worked fine locally). I will continue to troubleshoot! My attempts are in these commits: [49cda1c](https://github.com/kmartinet/SSARP/commit/49cda1c6b111a8e4870098db3c0de58814255986), [2e8afb8](https://github.com/kmartinet/SSARP/commit/2e8afb8ad76167baaa2e4bc98883dacb637041b1), [8bc92c5](https://github.com/kmartinet/SSARP/commit/8bc92c5106c424b84f2999b37c9c959a19df5b1f), and [f79401e](https://github.com/kmartinet/SSARP/commit/f79401e8649b875d5140247c6dcca93a97f84cb2)

**Examples response:**
([The below changes can be found in commit 7f29be1](https://github.com/kmartinet/SSARP/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
-  Examples that use a tree now access them with ```read.tree(system.file...))```
- Removed most of the `\dontrun` blocks in the examples, with the exception of
-- `get_key.R`: I left one call out of the `\dontrun{}` block, but kept the rest in to ensure that the GBIF API isn't being called too often in the examples
-- `create_SpAR.R`: I need to curate a better example dataset that creates a viable SpAR so that the examples don't rely on calling GBIF
-- `estimate_BAMM.R`: Same reason as above, and I need to do further BAMM testing to create a good example file

**General comments response:**
- Thank you so much for the complement about my package! I am currently working on writing more documentation about the use of `CoordinateCleaner` and `sars` in conjunction with `SSARP`.

**Dependencies response:**
- While I did update `get_key()` and `get_data()` as part of the changes I discussed here, I absolutely understand your point that it is better to not provide such wrapper functions. Instead, I will write helpful documentation about how to use `rgbif::occ_download()` in conjunction with `SSARP`.

**Naming response:**
([The below changes can be found in commit 7f29be1](https://github.com/kmartinet/SSARP/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- I am still considering other potential names for the package, but if I'm not inspired, I will simply call it `ssarp` in lowercase!
- I have changed all of the function names to follow `verb_object()` format. I recognize that the order is reverse from your `object_verb()` suggestion, but I find that `verb_object()` makes more sense for the function names I'd like to use.

**Documentation response:**
([The below changes can be found in commit 7f29be1](https://github.com/kmartinet/SSARP/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- All function names in the documentation are written as `package::function()` now
- I've added a reference to Magallon and Sanderson (2001) in `estimate_MS()`

**Style response:**
- Instead of using `@importFrom` and `@import`, the `package::function()` notation is used throughout the package now ([Commit 7f29be1](https://github.com/kmartinet/SSARP/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b)) with one exception:
-- In `find_land.R`, there is an interaction with the `maps` and `mapdata` packages when I call the `maps::map.where()` function that requires me to import both of these packages in order to avoid a note about dependencies ("Namespace in Imports field not imported from: 'mapdata' All declared Imports should be used."). I will continue to explore options here!

**Methodology and Etc response:**
- I will respond more in-depth when I've had a chance to fully address all of your comments, but I wanted to say thanks so much for drawing my attention to these issues!
- I was hoping to avoid spatial dependencies in `SSARP` since they can be difficult to install, but your suggestion to use a spatial dataframe for accessing island areas is an excellent idea that I think is worth the dependencies. I've been testing a few solutions and I am excited to share in the next version!
- I was originally using the `epm` package for estimating speciation rates using DR, but I removed the dependency due to `epm`'s spatial dependencies that `SSARP` was not using. With the addition of a new workflow for determining island area that involves spatial information, I will likely use the `epm` implementation of DR again.
- I wrote a short description of WKT format at the bottom of the `README`, but it's a confusing format and I should do a much better job describing it! I'm working on a separate vignette to help clear it up more.

## Partial response to @txm676
**General comments response**:
- Thank you so much for your kind words about my package!

**R checks response**:
- The leftover "figure" directory has been deleted.

**Model fitting and output response**:
-  I had not considered users providing islands with zero species, and this would absolutely make the model fitting functions error. I will work on catching these kinds of inputs!
- I have added `SSARP::get_presence_absence()`, which creates a presence_absence matrix when given a dataframe from `SSARP::find_areas()` ([Commit 7f29be1](https://github.com/kmartinet/SSARP/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b)). I will add a function for creating a standard island-species richness dataframe for my next response.

**Code response:**
- I have changed all of the function names to follow `verb_object()` format, which is hopefully more intuitive and consistent now. I will change the package name to `ssarp` in lowercase if I don't think of a different, more memorable name.
- While I did update `get_data()` as part of the changes I discussed here, I agree that it is an unnecessary function. Instead, I will write helpful documentation about how to use `rgbif::occ_download()` in conjunction with `SSARP`.

**Documentation response:**
- Thanks so much for reading the preprint of the `SSARP` manuscript! I really appreciate your comment that there is a lot of useful information there that should be added to the documentation. I am still working on a good balance for including that information across the package.
- I agree that there should be clearer instruction on how to build SARs for islands within an archipelago! I'm working on a vignette covering this topic, as well as how to properly encode a WKT polygon for use in `SSARP::get_data()` or `rgbif::occ_download()`

**Input data response:**
- ```Patton_Anolis_Trimmed.tree``` is now in ```inst/extdata/``` ([In commit 7f29be1](https://github.com/kmartinet/SSARP/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- The code for accessing it across the package has been updated to ```read.tree(system.file("extdata", "Patton_Anolis_Trimmed.tree", package = "SSARP"))``` with three important exceptions: `test-estimate_BAMM()`, `test-estimate_DR()`, and `test-estimate_MS()`. The test-coverage.yaml workflow kept failing due to it not being able to find the file (even though it worked fine locally). I will continue to troubleshoot! My attempts are in these commits: [49cda1c](https://github.com/kmartinet/SSARP/commit/49cda1c6b111a8e4870098db3c0de58814255986), [2e8afb8](https://github.com/kmartinet/SSARP/commit/2e8afb8ad76167baaa2e4bc98883dacb637041b1), [8bc92c5](https://github.com/kmartinet/SSARP/commit/8bc92c5106c424b84f2999b37c9c959a19df5b1f), and [f79401e](https://github.com/kmartinet/SSARP/commit/f79401e8649b875d5140247c6dcca93a97f84cb2)
- With my new approach of encouraging users to obtain data through `rgbif::occ_download()`, the max threshold of 100,000 occurrence records will not apply anymore, thankfully!

**Other points response:**
- I am working on restructuring the `README` to include more useful information per your suggestions, and I will include a link to the `sars` R package with information on how to efficiently use the two together. I take it as a huge complement that you are considering adding a link to `SSARP` to the `sars` `README`. Thanks so much! I agree that the packages nicely complement each other, and I am hopeful that users will find this to be true as well.
- Thank you for the comment related to the paper too! The idea I was hoping to convey with that part of the paper was that it has been generally difficult to conduct global comparisons, but I agree that it is unfair to say that they are rare.
