### 2025-07-15 82cf553
Documentation updates (pkgdown, speciation methods)
