+++ man/estimate_BAMM.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/estimate_BAMM.R
-\name{speciationBAMM}
-\alias{speciationBAMM}
+\name{estimate_BAMM}
+\alias{estimate_BAMM}
 \title{Get tip speciation rates from BAMM (Rabosky 2014) analysis}
 \usage{
-speciationBAMM(label_type = "binomial", occurrences, edata)
+estimate_BAMM(label_type = "binomial", occurrences, edata)
 }
 \arguments{
 \item{label_type}{Either "epithet" or "binomial" (default): describes the
@@ -50,5 +50,15 @@ edata <- BAMMtools::getEventData(phy = tree, eventdata = event_data)
 occ_speciation <- speciationBAMM(label_type = "epithet", 
                                  occurrences = areas , 
                                  edata = edata)
+}
+}
+\references{
+\itemize{
+\item Rabosky, D.L. (2014). Automatic Detection of Key Innovations, Rate Shifts,
+and Diversity-Dependence on Phylogenetic Trees. PLOS ONE, 9(2): e89543.
+\item Rabosky, D.L., Grundler, M., Anderson, C., Title, P., Shi, J.J.,
+Brown, J.W., Huang, H., & Larson, J.G. (2014), BAMMtools: an R package for
+the analysis of evolutionary dynamics on phylogenetic trees. Methods in
+Ecology and Evolution, 5: 701-707.
 }
 }
