+++ NAMESPACE
@@ -4,6 +4,7 @@ S3method(plot,SAR)
 S3method(print,SAR)
 export(create_SAR)
 export(create_SpAR)
+export(estimate_BAMM)
 export(estimate_DR)
 export(estimate_MS)
 export(find_areas)
@@ -15,7 +16,6 @@ export(get_presence_absence)
 export(get_sources)
 export(quick_create_SAR)
 export(remove_continents)
-export(speciationBAMM)
 import(mapdata)
 import(maps)
 importFrom(stringi,stri_unescape_unicode)

[…]
