### 2025-07-15 249af89
Removed typo in pkgdown.yml
