### 2025-07-15 c708a88
Added missing topics to pkgdown.yml
