+++ tests/testthat/test-estimate_DR.R
@@ -14,13 +14,13 @@ tree <- ape::read.tree("https://raw.githubusercontent.com/kmartinet/SSARP/refs/h
 # The estimate_DR function only really needs "Genus" and "Species" 
 #  columns to work, so a simple dataframe can be created for testing 
 #  (without real occurrence data)
-Species <- tree$tip.label
-Genus <- rep("Anolis", length(Species))
-occ_simple <- as.data.frame(cbind(Genus, Species))
+specificEpithet <- tree$tip.label
+genericName <- rep("Anolis", length(specificEpithet))
+occ_simple <- as.data.frame(cbind(genericName, specificEpithet))
 
 # Create a tree with binomial labels
 tree_b <- tree
-binom_names <- paste(Genus, Species, sep = " ")
+binom_names <- paste(genericName, specificEpithet, sep = " ")
 tree_b$tip.label <- binom_names
 
 # Matrix version of dataframe
