+++ pkgdown.yml
@@ -4,7 +4,8 @@ pkgdown_sha: ~
 articles:
   Create_SAR: Create_SAR.html
   Create_SpAR: Create_SpAR.html
-last_built: 2025-07-15T22:26Z
+  Get_Data: Get_Data.html
+last_built: 2025-07-17T23:41Z
 urls:
   reference: https://kmartinet.github.io/SSARP/reference
   article: https://kmartinet.github.io/SSARP/articles
