### 2025-07-18 0128566
Documentation fixes, removed unnecessary functions
Functions removed: get_data(), get_key(), quick_create_SAR()
