+++ R/remove_continents.R
@@ -5,8 +5,8 @@
 #' @param occs The dataframe that is returned by `SSARP::find_areas()`. I do not 
 #' recommend using a custom dataframe for this function because it references 
 #' areas given by the area database used in `SSARP::find_areas()`. If you must 
-#' use a custom dataframe, please ensure  that the landmass areas are in the 9th 
-#' column of the dataframe.
+#' use a custom dataframe, please ensure  that the landmass areas are in a
+#' column called "areas"
 #' @return A dataframe of the species name, island name, and island area 
 #' (without continents)
 #' @examples 
@@ -32,7 +32,7 @@ remove_continents <- function(occs){
   #  SSARP::find_areas()
   continents <- c(5.50e13, 3.04e13, 1.78e13, 2.47e13)
   for(i in seq(continents)){
-    exclude <- which(occs[,10] == continents[i])
+    exclude <- which(occs[,"areas"] == continents[i])
     if(length(exclude) != 0){
       occs <- occs[-exclude,]
     }
