### 2025-07-19 d8d19fe
Changed package name to 'ssarp', added test cases
