+++ R/find_areas.R
@@ -1,30 +1,30 @@
 #' Find areas of land masses.
-#' 
-#' Reference a dataset of island names and areas to find the areas of the land 
+#'
+#' Reference a dataset of island names and areas to find the areas of the land
 #' masses relevant to the taxon of interest.
-#' @param occs The dataframe that is returned by `SSARP::find_land()`. If using 
-#' a custom occurrence record dataframe, ensure that it has the following 
-#' columns: "acceptedScientificName", "genericName", "specificEpithet", 
-#' "decimalLongitude", "decimalLatitude", "First", "Second", "Third", 
-#' "datasetKey". The "datasetKey" column is important for GBIF records and 
-#' identifies the dataset to which the occurrence record belongs. Custom 
-#' dataframes without this style of data organization should fill the column 
+#' @param occs The dataframe that is returned by `ssarp::find_land()`. If using
+#' a custom occurrence record dataframe, ensure that it has the following
+#' columns: "acceptedScientificName", "genericName", "specificEpithet",
+#' "decimalLongitude", "decimalLatitude", "First", "Second", "Third",
+#' "datasetKey". The "datasetKey" column is important for GBIF records and
+#' identifies the dataset to which the occurrence record belongs. Custom
+#' dataframes without this style of data organization should fill the column
 #' with placeholder values.
-#' @param area_custom A dataframe including names of land masses and their 
-#' associated areas. This dataframe should be provided when the user would like 
-#' to bypass using the built-in database of island names and areas. Please 
-#' ensure that the custom dataframe includes the land mass's area in column 3 
+#' @param area_custom A dataframe including names of land masses and their
+#' associated areas. This dataframe should be provided when the user would like
+#' to bypass using the built-in database of island names and areas. Please
+#' ensure that the custom dataframe includes the land mass's area in column 3
 #' and the name in column 5. (Optional)
 #' @return A dataframe of the species name, island name, and island area
-#' @examples 
+#' @examples
 #' # The GBIF key for the Anolis genus is 8782549
-#' #  Obtained with: key <- get_key(query = "Anolis", rank = "genus")
-#' key <- 8782549
-#' # Read in example dataset obtained through:
-#' #  dat <- get_data(key = key, limit = 100)
+#' # Read in example dataset filtered from:
+#' #  dat <- rgbif::occ_search(taxonKey = 8782549, 
+#' #                           hasCoordinate = TRUE,
+#' #                           limit = 10000)
 #' dat <- read.csv(system.file("extdata",
-#'                             "SSARP_Example_Dat.csv",
-#'                             package = "SSARP"))
+#'                             "ssarp_Example_Dat.csv",
+#'                             package = "ssarp"))
 #' occs <- find_land(occurrences = dat)
 #' areas <- find_areas(occs = occs)
 #' @export
@@ -32,136 +32,164 @@
 find_areas <- function(occs, area_custom = NULL) {
   # checkmate input verification
   checkmate::assertDataFrame(occs)
-  
+  checkmate::testSubset(
+    c(
+      "acceptedScientificName",
+      "genericName",
+      "specificEpithet",
+      "decimalLongitude",
+      "decimalLatitude",
+      "First",
+      "Second",
+      "Third",
+      "datasetKey"
+    ),
+    names(occs)
+  )
+  # Ensure columns are correct type
[…]
