+++ R/get_presence_absence.R
@@ -1,73 +1,93 @@
 #' Create a presence-absence dataframe for a given occurrence record dataframe
-#' 
-#' Use a dataframe output by `SSARP::find_areas()` to determine which species 
-#' occur on specific islands by creating a presence-absence dataframe. A 1 
+#'
+#' Use a dataframe output by `ssarp::find_areas()` to determine which species
+#' occur on specific islands by creating a presence-absence dataframe. A 1
 #' represents presence and a 0 represents absence.
-#' @param occs The dataframe output by `SSARP::find_areas()`, or if using a 
+#' @param occs The dataframe output by `ssarp::find_areas()`, or if using a
 #' custom dataframe, ensure that it has the following named columns:
 #' - "areas" containing the areas associated with the land masses of interest
-#' - "Species" containing the names of the species living on those islands
-#' - "First" containing locality information. In the SSARP workflow, this 
+#' - "specificEpithet" containing the names of the species living on those
+#' islands
+#' - "First" containing locality information. In the ssarp workflow, this
 #' column contains the country name
-#' - "Second" containing locality information. In the SSARP workflow, this 
+#' - "Second" containing locality information. In the ssarp workflow, this
 #' column contains a province or island name
-#' - "Third" containing locality information. In the SSARP workflow, this 
-#' column contains the island name if the 7th column does not contain the 
+#' - "Third" containing locality information. In the ssarp workflow, this
+#' column contains the island name if the 7th column does not contain the
 #' island name
-#' @return A dataframe with a row for each island in the given occurrence 
-#' record dataframe and a column for each species. Within each species column, 
-#' a 1 represents the presence of that species on the island corresponding to 
-#' the given row, and a 0 represents the absence of that species on the island 
+#' @return A dataframe with a row for each island in the given occurrence
+#' record dataframe and a column for each species. Within each species column,
+#' a 1 represents the presence of that species on the island corresponding to
+#' the given row, and a 0 represents the absence of that species on the island
 #' corresponding to the given row.
-#' @examples 
+#' @examples
 #' # The GBIF key for the Anolis genus is 8782549
-#' #  Obtained with: key <- get_key(query = "Anolis", rank = "genus")
-#' key <- 8782549
-#' # Read in example dataset obtained through:
-#' #  dat <- get_data(key = key, limit = 100)
+#' # Read in example dataset filtered from:
+#' #  dat <- rgbif::occ_search(taxonKey = 8782549, 
+#' #                           hasCoordinate = TRUE,
+#' #                           limit = 10000)
 #' dat <- read.csv(system.file("extdata",
-#'                             "SSARP_Example_Dat.csv",
-#'                             package = "SSARP"))
+#'                             "ssarp_Example_Dat.csv",
+#'                             package = "ssarp"))
 #' land <- find_land(occurrences = dat)
 #' areas <- find_areas(occs = land)
 #' pres_abs <- get_presence_absence(areas)
-#' 
+#'
 #' @export
 
 get_presence_absence <- function(occs) {
+  # Checkmate input validation
+  checkmate::assertDataFrame(occs)
+  checkmate::testSubset(
+    c("specificEpithet", "areas", "First", "Second", "Third"),
+    names(occs)
+  )
+  # Ensure columns are correct type
+  checkmate::assertCharacter(occs$specificEpithet)
+  checkmate::assertNumeric(occs$areas)
+  checkmate::assertCharacter(occs$First)
+  checkmate::assertCharacter(occs$Second)
+  checkmate::assertCharacter(occs$Third)
+
[…]
