### 2025-07-19 d0df42d
Added plotting option to create_SAR/SpAR, internet handling
