+++ R/create_SpAR.R
@@ -9,6 +9,8 @@
 #' has the following columns: areas, rate
 #' @param npsi The maximum number of breakpoints to estimate for model
 #' selection.  Default: 1
+#' @param visualize (boolean) Whether the plot should be displayed when the 
+#' function is called. Default: FALSE
 #' @return A list of 3 including: the summary output, the segmented regression
 #' object, and the aggregated dataframe used to create the plot
 #' @examples
@@ -32,11 +34,14 @@
 #'                               label_type = "epithet",
 #'                               occurrences = areas)
 #' 
-#' seg <- create_SpAR(occurrences = occ_speciation, npsi = 1)
+#' seg <- create_SpAR(occurrences = occ_speciation, 
+#'                    npsi = 1,
+#'                    visualize = FALSE)
+#' plot(seg)
 #' summary <- seg[1]
 #' @export
 
-create_SpAR <- function(occurrences, npsi = 1) {
+create_SpAR <- function(occurrences, npsi = 1, visualize = FALSE) {
   # Checkmate input validation
   checkmate::assertDataFrame(occurrences)
   checkmate::assertNumeric(npsi)
@@ -109,16 +114,18 @@ create_SpAR <- function(occurrences, npsi = 1) {
 
   # If the min_score is index 1, then the best-fit model is linear
   if (min_score == 1) {
-    plot(
-      dat,
-      xlim = c(x_min, (x_max + 0.5)),
-      ylim = c(y_min, y_buff),
-      ylab = "Log Speciation Rate",
-      xlab = expression(paste("Log Island Area (", "m"^"2", ")")),
-      main = "Speciation-Area Relationship",
-      pch = 16
-    )
-    graphics::abline(linear)
+    if(visualize){
+      plot(
+        dat,
+        xlim = c(x_min, (x_max + 0.5)),
+        ylim = c(y_min, y_buff),
+        ylab = "Log Speciation Rate",
+        xlab = expression(paste("Log Island Area (", "m"^"2", ")")),
+        main = "Speciation-Area Relationship",
+        pch = 16
+      )
+      graphics::abline(linear)
+    }
 
     summary_line <- summary(linear)
 
@@ -146,17 +153,19 @@ create_SpAR <- function(occurrences, npsi = 1) {
     )
 
     # Plot the breakpoint regression line
-    plot(
-      seg,
-      rug = FALSE,
-      xlim = c(x_min, (x_max + 0.5)),
-      ylim = c(y_min, y_buff),
-      ylab = "Log Speciation Rate",
-      xlab = expression(paste("Log Island Area (", "m"^"2", ")")),
-      main = "Speciation-Area Relationship"
-    )
-    # Add the points
-    graphics::points(dat$x, dat$y, pch = 19)
+    if(visualize){
+      plot(
+        seg,
+        rug = FALSE,
+        xlim = c(x_min, (x_max + 0.5)),
+        ylim = c(y_min, y_buff),
+        ylab = "Log Speciation Rate",
+        xlab = expression(paste("Log Island Area (", "m"^"2", ")")),
+        main = "Speciation-Area Relationship"
+      )
+      # Add the points
+      graphics::points(dat$x, dat$y, pch = 19)
+    }
 
     # Save the summary as an object to add to the result list
     summary_seg <- summary(seg)
@@ -187,7 +196,9 @@ create_SpAR <- function(occurrences, npsi = 1) {
 
     # Plot defaults to multiple outputs when npsi > 1, so my npsi = 1 plot
     #  doesn't apply
-    plot(seg)
+    if(visualize){
+      plot(seg)
+    }
 
     # Save the summary as an object to add to the result list
     summary_seg <- summary(seg)
