### 2025-07-19 379bdc1
Added get_richness() and spatial inputs to find_areas()
