+++ R/find_areas.R
@@ -1,7 +1,24 @@
 #' Find areas of land masses.
 #'
-#' Reference a dataset of island names and areas to find the areas of the land
-#' masses relevant to the taxon of interest.
+#' Find the areas of the land masses relevant to the taxon of interest with two
+#' options: a database of island names and areas, or a user-provided shapefile. 
+#' 
+#' The first method is to reference a built-in dataset of island names 
+#' and areas to find the areas of the landmasses relevant to the taxon of 
+#' interest. The user may also decide to input their own custom dataframe 
+#' including names of relevant land masses and their associated areas to 
+#' bypass using *ssarp*'s built-in dataset.
+#' 
+#' The second method is to reference a user-supplied shapefile containing
+#' spatial information for the landmasses of interest in order to determine
+#' their areas. 
+#' 
+#' While the word "landmasses" was used heavily in this documentation, users
+#' supplying their own custom area dataframe or shapefile are encouraged to
+#' use this function in the *ssarp* workflow to create species- and speciation-
+#' area relationships for island-like systems such as lakes, fragmented habitat,
+#' and mountain peaks.
+#' 
 #' @param occs The dataframe that is returned by `ssarp::find_land()`. If using
 #' a custom occurrence record dataframe, ensure that it has the following
 #' columns: "acceptedScientificName", "genericName", "specificEpithet",
@@ -13,8 +30,15 @@
 #' @param area_custom A dataframe including names of land masses and their
 #' associated areas. This dataframe should be provided when the user would like
 #' to bypass using the built-in database of island names and areas. Please
-#' ensure that the custom dataframe includes the land mass's area in column 3
-#' and the name in column 5. (Optional)
+#' ensure that the custom dataframe includes the land mass's area in a column
+#' called "AREA" and the name in a column called "Name". (Optional)
+#' @param shapefile A shapefile (.shp) containing spatial information for 
+#' the geographic locations of interest. (Optional)
+#' @param names If the user would like to restrict which polygons in the 
+#' shapefile are included in the returned occurrence record dataframe, they can 
+#' be specified here as a vector. If the user does not provide a vector, all of 
+#' the non-NA names in the shapefile will be included 
+#' (as found in shapefile$name). (Optional)
 #' @return A dataframe of the species name, island name, and island area
 #' @examples
 #' # The GBIF key for the Anolis genus is 8782549
@@ -29,7 +53,8 @@
 #' areas <- find_areas(occs = occs)
 #' @export
 
-find_areas <- function(occs, area_custom = NULL) {
+find_areas <- function(occs, area_custom = NULL, 
+                       shapefile = NULL, names = NULL) {
   # checkmate input verification
   checkmate::assertDataFrame(occs)
   checkmate::testSubset(
@@ -58,138 +83,180 @@ find_areas <- function(occs, area_custom = NULL) {
   # Not checking datasetKey because it is not relevant to the code and can be
   #  any type, really
 
-  # Remove rows where First, Second, and Third are all NA
-  # Create vector to hold row numbers
-  minus <- rep(NA, nrow(occs))
-  # Loop through dataframe
-  for (i in seq_len(nrow(occs))) {
-    if (nrow(occs) == 0) {
-      cli::cli_alert_warning("No data in occurrence record dataframe")
-      break
-    }
-    if (
-      is.na(occs[i, "Third"]) &&
-        is.na(occs[i, "Second"]) &&
-        is.na(occs[i, "First"])
-    ) {
[…]
