+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Create_SAR: Create_SAR.html
   Create_SpAR: Create_SpAR.html
   Get_Data: Get_Data.html
-last_built: 2025-07-19T04:05Z
+last_built: 2025-07-19T16:55Z
 urls:
   reference: https://kmartinet.github.io/ssarp/reference
   article: https://kmartinet.github.io/ssarp/articles
