+++ README.md
@@ -15,44 +15,50 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 Review](https://badges.ropensci.org/685_status.svg)](https://github.com/ropensci/software-review/issues/685)
 <!-- badges: end -->
 
-SSARP (Species-/Speciation-Area Relationship Projector) is an R package
-that provides a suite of functions to help users **create speciation-
-and species-area relationships for island-dwelling taxa** using
-occurrence data from GBIF (Global Biodiversity Information Facility)
-accessed via GBIF’s API or the user’s own occurrence data.
+`ssarp` (Species-/Speciation-Area Relationship Projector) is an R package that 
+provides a suite of functions to help users **create speciation- and species-
+area relationships for island-dwelling taxa** using occurrence data from GBIF
+(Global Biodiversity Information Facility) or the user's own occurrence data. 
 
-Species- and speciation-area relationships are useful for determining
-whether biodiversity is distributed in an unusual way in a given island
-system (e.g., species richness decreasing as islands get larger instead
-of increasing). An unusual trend in these relationships may be
-indicative of habitat loss or a negative impact of non-native species.
+Species- and speciation-area relationships are useful for determining whether
+biodiversity is distributed in an unusual way in a given island system 
+(e.g., species richness decreasing as islands get larger instead of 
+increasing). An unusual trend in these relationships may be indicative of 
+habitat loss or a negative impact of non-native species.
 
-Please find the [bioRxiv preprint of the manuscript associated with
-*SSARP*
-here!](https://www.biorxiv.org/content/10.1101/2024.12.31.630948v1)
+Please find the [bioRxiv preprint of the manuscript associated with *ssarp* here!](https://www.biorxiv.org/content/10.1101/2024.12.31.630948v1)
 
 ## Installation
 
-To install *SSARP*, use the “install_github” function from the
-*devtools* package:
+To install *ssarp*, use the "install_github" function from the *devtools* package:
 
 ``` r
 # install.packages("devtools")
 library(devtools)
 
-install_github("kmartinet/SSARP")
+install_github("kmartinet/ssarp")
 ```
+
+## Suppressing Messages
+
+To suppress all messages output by *ssarp*, run
+
+``` r
+options(ssarp.silent = TRUE)
+```
+
+before using any of *ssarp*'s functions.
 
 ## Example: Creating a Species-Area Relationship
 
 A species-area relationship (SAR) visualizes the relationship between
 species richness (the number of species) and the area of the land mass
-on which the species live. This brief example covers the *SSARP*
+on which the species live. This brief example covers the *ssarp*
 workflow for creating a SAR, and more detailed explanations of the code
 and methods can be found [in the Articles on the SSARP pkgdown
-website](https://kmartinet.github.io/SSARP/index.html).
+website](https://kmartinet.github.io/ssarp/index.html).
 
-In order to construct a species-area relationship with *SSARP*, we will:
+In order to construct a species-area relationship with *ssarp*, we will:
 
 - Gather occurrence data from GBIF
 - Filter out invalid occurrence records
@@ -61,32 +67,39 @@ In order to construct a species-area relationship with *SSARP*, we will:
 
 #### Step 1: Gather occurrence data from GBIF
 
-In this step, we will find the unique identifying key associated with a
+In this step, we will find the unique identifying key associated with a 
 taxon of interest (the lizard genus *Anolis* in this case) and use that
-key to access occurrence points for that taxon from GBIF. We will
+key to access occurrence points for that taxon from GBIF. We will 
 restrict the returned occurrence records to the Caribbean islands using
-the “geometry” parameter in the “getData” function.
[…]
