+++ R/create_SAR.R
@@ -3,6 +3,11 @@
 #' Use segmented regression to create a species-area relationship (SAR) plot.
 #' The X axis represents log(island area) and the Y axis represents log(number
 #' of species)
+#' 
+#' If the user would prefer to create their own plot of the 
+#' `ssarp::create_SAR()` output, the `aggDF` element of the returned list
+#' includes the raw points from the plot created here. They can be accessed
+#' as demonstrated in the Examples section.
 #' @param occurrences The dataframe output by `ssarp::find_areas()` (or if
 #' using  a custom dataframe, ensure that it has the following columns:
 #' specificEpithet, areas)
@@ -10,8 +15,9 @@
 #' selection.  Default: 1
 #' @param visualize (boolean) Whether the plot should be displayed when the 
 #' function is called. Default: FALSE
-#' @return A list of 3 including: the summary output, the segmented regression
-#' object, and the aggregated dataframe used to create the plot
+#' @return A list of 4 including: the summary output, the regression
+#' object, the aggregated dataframe used to create the plot, and the AIC scores
+#' used in model selection
 #' @examples
 #' # The GBIF key for the Anolis genus is 8782549
 #' # Read in example dataset filtered from:
@@ -28,7 +34,9 @@
 #'                   npsi = 1,
 #'                   visualize = FALSE)
 #' plot(seg)
-#' summary <- seg[1]
+#' summary <- seg$summary
+#' model_object <- seg$segObj
+#' points <- seg$aggDF
 #' @export
 
 create_SAR <- function(occurrences, npsi = 1, visualize = FALSE) {
