### 2025-07-19 fa01fd3
Documentation edits, update NEWS
