+++ R/create_SpAR.R
@@ -3,18 +3,21 @@
 #' Use segmented regression to create a speciation-area relationship plot. The
 #' X axis represents log(island area) and the Y axis represents
 #' log(speciation rate)
-#' 
-#' If the user would prefer to create their own plot of the 
+#'
+#' If the user would prefer to create their own plot of the
 #' `ssarp::create_SpAR()` output, the `aggDF` element of the returned list
 #' includes the raw points from the plot created here. They can be accessed
 #' as demonstrated in the Examples section.
+#'
+#' More information about the three methods for estimating speciation rate
+#' included in `ssarp` can be found in [ssarp's SpAR vignette](https://kmartinet.github.io/ssarp/articles/Create_SpAR.html).
 #' @param occurrences The dataframe output by one of ssarp's speciation
 #' methods (`ssarp::estimate_BAMM()`, `ssarp::estimate_DR()`,
 #' `ssarp::estimate_MS()`), or if using a custom dataframe, ensure that it
 #' has the following columns: areas, rate
 #' @param npsi The maximum number of breakpoints to estimate for model
 #' selection.  Default: 1
-#' @param visualize (boolean) Whether the plot should be displayed when the 
+#' @param visualize (boolean) Whether the plot should be displayed when the
 #' function is called. Default: FALSE
 #' @return A list of 4 including: the summary output, the regression
 #' object, the aggregated dataframe used to create the plot, and the AIC scores
@@ -22,7 +25,7 @@
 #' @examples
 #' # The GBIF key for the Anolis genus is 8782549
 #' # Read in example dataset filtered from:
-#' #  dat <- rgbif::occ_search(taxonKey = 8782549, 
+#' #  dat <- rgbif::occ_search(taxonKey = 8782549,
 #' #                           hasCoordinate = TRUE,
 #' #                           limit = 10000)
 #' dat <- read.csv(system.file("extdata",
@@ -30,7 +33,7 @@
 #'                             package = "ssarp"))
 #' land <- find_land(occurrences = dat)
 #' areas <- find_areas(occs = land)
-#' 
+#'
 #' # Read tree from Patton et al. (2021), trimmed to Caribbean species
 #' tree <- ape::read.tree(system.file("extdata",
 #'                                    "Patton_Anolis_trimmed.tree",
@@ -39,8 +42,8 @@
 #' occ_speciation <- estimate_MS(tree = tree,
 #'                               label_type = "epithet",
 #'                               occurrences = areas)
-#' 
-#' seg <- create_SpAR(occurrences = occ_speciation, 
+#'
+#' seg <- create_SpAR(occurrences = occ_speciation,
 #'                    npsi = 1,
 #'                    visualize = FALSE)
 #' plot(seg)
@@ -122,7 +125,7 @@ create_SpAR <- function(occurrences, npsi = 1, visualize = FALSE) {
 
   # If the min_score is index 1, then the best-fit model is linear
   if (min_score == 1) {
-    if(visualize){
+    if (visualize) {
       plot(
         dat,
         xlim = c(x_min, (x_max + 0.5)),
@@ -161,7 +164,7 @@ create_SpAR <- function(occurrences, npsi = 1, visualize = FALSE) {
     )
 
     # Plot the breakpoint regression line
-    if(visualize){
+    if (visualize) {
       plot(
         seg,
         rug = FALSE,
@@ -204,7 +207,7 @@ create_SpAR <- function(occurrences, npsi = 1, visualize = FALSE) {
 
     # Plot defaults to multiple outputs when npsi > 1, so my npsi = 1 plot
     #  doesn't apply
-    if(visualize){
+    if (visualize) {
       plot(seg)
     }
 
