+++ R/get_sources.R
@@ -17,7 +17,7 @@
 #' @examples
 #' # The GBIF key for the Anolis genus is 8782549
 #' # Read in example dataset filtered from:
-#' #  dat <- rgbif::occ_search(taxonKey = 8782549, 
+#' #  dat <- rgbif::occ_search(taxonKey = 8782549,
 #' #                           hasCoordinate = TRUE,
 #' #                           limit = 10000)
 #' dat <- read.csv(system.file("extdata",
@@ -43,7 +43,7 @@ get_sources <- function(occs) {
 
     return(count_df)
   } else {
-    if(!getOption("ssarp.silent", FALSE)){
+    if (!getOption("ssarp.silent", FALSE)) {
       cli::cli_alert_warning("datasetKey column not found.")
     }
     return(NULL)
