### 2025-07-20 #685 @kmartinet
Hello again @jhollist , @joelnitta , and @txm676 ! Thank you so much for your patience as I finished my response to your reviews, and thanks so much again for your thoughtful comments.

Instead of simply continuing from my response on the 4th, I will include all of my responses here so you don't have to look in two different places.

[The NEWS.md update for these changes](https://github.com/kmartinet/ssarp/blob/main/NEWS.md)

## Response to @joelnitta
**Vignettes response:**
- ```Patton_Anolis_Trimmed.tree``` is now in ```inst/extdata/``` ([In commit 7f29be1](https://github.com/kmartinet/ssarp/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- The code for accessing it across the package has been updated to ```read.tree(system.file("extdata", "Patton_Anolis_Trimmed.tree", package = "SSARP"))``` with three important exceptions: `test-estimate_BAMM()`, `test-estimate_DR()`, and `test-estimate_MS()`. The test-coverage.yaml workflow kept failing due to it not being able to find the file (even though it worked fine locally). My attempts are in these commits: [49cda1c](https://github.com/kmartinet/ssarp/commit/49cda1c6b111a8e4870098db3c0de58814255986), [2e8afb8](https://github.com/kmartinet/ssarp/commit/2e8afb8ad76167baaa2e4bc98883dacb637041b1), [8bc92c5](https://github.com/kmartinet/ssarp/commit/8bc92c5106c424b84f2999b37c9c959a19df5b1f), and [f79401e](https://github.com/kmartinet/ssarp/commit/f79401e8649b875d5140247c6dcca93a97f84cb2)

The use of ```inst/extdata/``` at all in examples and test files has created intermittent problems. In the most recent commit, R-CMD-check is failing because the Ubuntu distributions cannot find ```extdata```. It works for Windows and Macos, and it has also worked with Ubuntu in previous commits, so I am at a loss about what is causing the issues. Do you have any advice for how to proceed? I considered wrapping each instance of using ```extdata``` in a statement that will skip it if the OS believes the file does not exist, but I wanted to ask for guidance before doing that.

**Examples response:**
- None of the examples are wrapped in ```\dontrun{}``` now
-  Examples that use a tree now access them with ```read.tree(system.file...))```  ([commit 7f29be1](https://github.com/kmartinet/ssarp/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- ```inst/extdata/ssarp_Example_Dat.csv``` is now useful for SpAR examples as well ([commit 672a75f](https://github.com/kmartinet/ssarp/commit/672a75f03f3deed84715c7ed9b0c9cad04a0d6f4))
- I added a new example file, ```inst/extdata/event_data_Patton_Anolis.txt```, for the example in ```ssarp::speciation_BAMM()``` ([commit d8d19fe](https://github.com/kmartinet/ssarp/commit/d8d19fedc5542d91a30844af0952c7281fc85d23))

**Automated tests response:**
- New tests for calculations performed by ```ssarp::create_SAR()```, ```ssarp::creaate_SpAR()```, ```ssarp::estimate_BAMM()```, ```ssarp::estimate_DR()```, and ```ssarp::estimate_MS()``` have been written. ([commit d8d19fe](https://github.com/kmartinet/ssarp/commit/d8d19fedc5542d91a30844af0952c7281fc85d23))

**General comments response:**
- Thank you so much for the complement about my package!
- I have removed the ```SSARP::getData()``` function and replaced it with [a vignette describing how to gather occurrence records from GBIF](https://kmartinet.github.io/ssarp/articles/Get_Data.html).
- I have removed the ```SSARP::quickSARP()``` function. In addition, the [new occurrence record vignette includes a section](https://kmartinet.github.io/ssarp/articles/Get_Data.html#curating-gbif-occurrence-records) on using ```CoordinateCleaner``` to curate data.

**Dependencies response:**
- I have removed the ```SSARP::getData()``` and ```SSARP::getKey()``` functions, and replaced them with [the vignette describing how to gather occurrence records from GBIF](https://kmartinet.github.io/ssarp/articles/Get_Data.html). The vignette encourages the use of ```rgbif::occ_download()```, but still includes information for how to use ```rgbif::occ_search()``` in conjunction with ```ssarp::get_sources()``` just in case that is helpful for users.

**Output format response:**
- I added text and code in the Examples for ```ssarp::create_SAR()``` and ```ssarp::create_SpAR``` to help make the metadata more easily accessible to users. ([commit b8399e6](https://github.com/kmartinet/ssarp/commit/b8399e61273f81f9daaa86cacb78df602014fb94))
- With the removal of ```SSARP::getData()```, there were no more uses of class ```tibble``` in the package. I chose to stick with class ```data.frame``` for my functions because I prefer to lean into ```base``` R whenever possible for basic functionality.
- Column names have been standardized across the package to follow the naming conventions of ```rgbif``` ([commit 672a75f](https://github.com/kmartinet/ssarp/commit/672a75f03f3deed84715c7ed9b0c9cad04a0d6f4))
- Plot printing in ```ssarp::create_SAR()``` and ```ssarp::create_SpAR()``` now only occurs when a user explicitly requests it ([commit d0df42d](https://github.com/kmartinet/ssarp/commit/d0df42da2daf4a5195c3929b0cc17f0b69e1de03))

**Naming response:**
- I didn't think of a new name, so the package name is now ```ssarp``` in all lowercase! ([commit d8d19fe](https://github.com/kmartinet/ssarp/commit/d8d19fedc5542d91a30844af0952c7281fc85d23))
- I have changed all of the function names to follow `verb_object()` format. I recognize that the order is reverse from your `object_verb()` suggestion, but I find that `verb_object()` makes more sense for the function names I'd like to use. ([commit 7f29be1](https://github.com/kmartinet/ssarp/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))

**Documentation response:**
- All function names in the documentation are written as `package::function()` now ([commit 7f29be1](https://github.com/kmartinet/ssarp/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- I've added a reference to Magallon and Sanderson (2001) in `ssarp::estimate_MS()` ([commit 7f29be1](https://github.com/kmartinet/ssarp/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- Columns are now referred to by name in `ssarp::find_areas()`, and elsewhere in the package ([commit 379bdc1](https://github.com/kmartinet/ssarp/commit/379bdc102f3949bcf83519be763c8eefea3a6930))

**UI response:**
- Please correct me if I'm misunderstanding, but is it sufficient to let users know that they can opt out of messages using `options(ssarp.silent = TRUE)`? I've added that functionality and added a line to the README explaining it. ([commit b8399e6](https://github.com/kmartinet/ssarp/commit/b8399e61273f81f9daaa86cacb78df602014fb94))

**Style response:**
- I installed the `air` formatter and it wasn't functioning properly with RStudio, but I was able to format using VS Code. I went through all of the files in `ssarp/R/` and applied the `air` formatter.
- Instead of using `@importFrom` and `@import`, the `package::function()` notation is used throughout the package now ([Commit 7f29be1](https://github.com/kmartinet/ssarp/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b)) with one exception:
- - In `find_land.R`, there is an interaction with the `maps` and `mapdata` packages when I call the `maps::map.where()` function that requires me to import both of these packages in order to avoid a note about dependencies ("Namespace in Imports field not imported from: 'mapdata' All declared Imports should be used."). I will continue to explore options here!

**Input checks response:**
- I've added `checkmate` input validation to determine whether certain columns are in the data, and that those columns are of a particular type. ([commit 379bdc1](https://github.com/kmartinet/ssarp/commit/379bdc102f3949bcf83519be763c8eefea3a6930))

**README response:**
- I've added the citation to the README.

**Website response:**
- The [Reference section of the ssarp website](https://kmartinet.github.io/ssarp/reference/index.html) now has the functions listed in groups.

**Tests response:**
- Tests that require an internet connection are now skipped if there is no internet connection ([commit d0df42d](https://github.com/kmartinet/ssarp/commit/d0df42da2daf4a5195c3929b0cc17f0b69e1de03))
- `test-estimate_BAMM.R` still does not use `BAMMtools` directly because of the `extdata` issue described at the beginning of this response. I will change this once we figure out how to fix the `extdata` issue!
- I have not yet implemented snapshot tests for plots because I was hopeful that the new tests covering the calculation of slopes and breakpoints for `ssarp::create_SAR()` and `ssarp::create_SpAR()` would cover this aspect. If not, I will work on snapshot tests for plots!

**Methodology response:**
- I appreciate the suggestion to check for spatial autocorrelation. I recognize that the nonindependence of islands in archipelagos is an incredibly important aspect that contributes to the shape of SARs and SpARs, but I ultimately chose not to include methodology to address spatial autocorrelation because researchers typically treat these effects as part of the broader biogeographic processes that are captured by SARs and SpARs (Triantis et al. 2012). That being said, if you believe that this methodology would make `ssarp` a much stronger addition to the island biogeography field, then I will gladly add functionality for checking spatial autocorrelation (potentially from `spdep`)!
- While occurrence data is used in `ssarp::create_SpAR()`, the `aggregate` function ensures that each species and its rate is only accounted for once per island (or at least, that's my understanding). I agree that abundance weighting would be a bad idea here!
- The methodology is different for `ssarp::estimate_MS()` because of the way the crown group diversification rate is calculated (Equation 4, Magallόn and Sanderson 2001). This diversification rate is meant to be calculated for a clade, and not for tips. DR and BAMM estimate tip speciation rates, so their methodology is more straightforward in that regard.

**Etc response:**
- The original message in `ssarp::find_land()` is now an error.
- Before I commit to using it, I wanted to ask if your API testing suggestions would be best addressed with [httptest](https://cran.r-project.org/web/packages/httptest/vignettes/httptest.html)? I am not experienced with writing responsible API calls in software, so I would appreciate your guidance!
- Columns are referred to by name instead of hard-coding position throughout the package now
- There is a buffer around the points in the plots created in both `ssarp::create_SAR()` and `ssarp::create_SpAR()`. The buffer in `create_SpAR()` is based on a percentage of the value, instead of a hardcoded 0.5 as in `create_SAR`, because speciation rates are typically very small and the same buffer would be too large for SpARs. I added a buffer in the first place because I noticed that some points would be cut off without one.
- Thank you for the correction in `ssarp::estimate_DR`!
- I was originally using the `epm` package for estimating speciation rates using DR, but I removed the dependency due to `epm`'s spatial dependencies that `ssarp` was not using at the time. I tested the current function (which is also associated with a published repository) extensively against `epm`'s DR function and I can confidently say that they return the same values. Would certain tests help boost confidence in the current iteration of `ssarp::estimate_DR`, or do you think it's best to use `epm` again instead?
- I will include a lookup table for tree inputs for `speciation_MS`, `speciation_BAMM`, and `speciation_DR` in my next response! I need to conduct more testing.
- The new (vignette covering gathering occurrence records from GBIF)[https://kmartinet.github.io/ssarp/articles/Get_Data.html#restricting-occurrence-records-using-geometry] includes a section with detailed instructions on how to create a WKT polygon and use it with `rgbif`.

**Response related to using spatial methods for obtaining island areas**
I wanted to create a new section to discuss this aspect because there is a lot of content here. First, thank you so much for your detailed thoughts about the problems with using the `island_areas` dataframe to obtain island areas!

I was really hopeful that using `rnaturalearth` would solve the problem, but their polygons are not granular enough for `ssarp`'s needs. Querying a combined dataset including `land` and `minor_islands` would return the area for the entire country (including the mainland) even for well-studied islands such as Santa Cruz and Isabela in the Galápagos Islands. I also tried `rnaturalearthhires` with the same result.

I also tried the `osmdata` package to attempt to use OpenStreetMap data directly to get island areas, but the required downloads were too large to include with `ssarp`.

With these failures, `ssarp` still uses the `island_areas` dataframe to access island areas. However, I've added functionality to `ssarp::find_areas()` that allows users to input their own shapefiles for use in gathering island areas. ([commit 379bdc1](https://github.com/kmartinet/ssarp/commit/379bdc102f3949bcf83519be763c8eefea3a6930#diff-71dfac25c8b89e565ea6892a266daa395ff29cc5f50245c7a007f7bd1b295f63))

I have not been able to attempt this idea yet, but I was also considering re-making the `island_areas` dataframe to include a column that categorizes each island by hemisphere (North, South, East, and West). There might still be multiple islands with the same name in the same hemispheres, but I thought this might help for users who do not have shapefiles.

**References**
- Magallόn, S. & Sanderson, M.J. (2001). Absolute Diversification Rates in Angiosperm Clades. Evolution, 55(9): 1762-1780.
- Triantis, K.A., Guilhaumon, F. and Whittaker, R.J. (2012), The island species–area relationship: biology and statistics. Journal of Biogeography, 39: 215-231. https://doi.org/10.1111/j.1365-2699.2011.02652.x

## Response to @txm676
**Automated tests response:**
- New tests for calculations performed by ```ssarp::create_SAR()```, ```ssarp::creaate_SpAR()```, ```ssarp::estimate_BAMM()```, ```ssarp::estimate_DR()```, and ```ssarp::estimate_MS()``` have been written. ([commit d8d19fe](https://github.com/kmartinet/ssarp/commit/d8d19fedc5542d91a30844af0952c7281fc85d23))

**General comments response:**
- Thank you so much for your kind words about my package!

**R checks response:**
- The leftover "figure" directory has been deleted.

**Model fitting and output response:**
- I would prefer to only return the best-fit model for `create_SAR` and `create_SpAR` because I like the simplicity. The AIC scores for all tested models are returned, so users may decide to re-run the analysis with different npsi if they would like to see what those models look like!
- I did some more testing, and if the user provides islands with zero species, the `agg` call in `create_SAR` and `create_SpAR` will remove these records! So the model fitting functions should work no matter the input.
- I have added `ssarp::get_presence_absence()`, which creates a presence_absence matrix when given a dataframe from `SSARP::find_areas()` ([Commit 7f29be1](https://github.com/kmartinet/ssarp/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b)). I have also added `ssarp::get_richness()`, which creates a standard species richness dataframe that can be easily input into `sars` for additional model creation ([commit 379bdc1](https://github.com/kmartinet/ssarp/commit/379bdc102f3949bcf83519be763c8eefea3a6930#diff-b2206eb8aff6428cead3476c4f556a679653d3435dc0db0ab54dc6eaaa2837f7))

**Code response:**
- I used the `air` formatter to standardize the code style across the package.
- The package name is now `ssarp` in lowercase ([commit d8d19fe](https://github.com/kmartinet/ssarp/commit/d8d19fedc5542d91a30844af0952c7281fc85d23)) and the function names have all been renamed to follow `verb_object()` format, which is hopefully more intuitive and consistent now ([commit 7f29be1](https://github.com/kmartinet/ssarp/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- Returning the plot is now an option in `create_SAR` and `create_SpAR`, and the default is to not return the plot ([commit d0df42d](https://github.com/kmartinet/ssarp/commit/d0df42da2daf4a5195c3929b0cc17f0b69e1de03))
- I definitely appreciate your comment about redundancy! I prefer to have separate methods for building the SAR and SpAR because I like being able to delineate between the two in vignettes and other documentation. `create_SpAR` also has a different method for adding a buffer to the plot due to the small values it plots.
- `SSARP::getData()` has been deleted in favor of adding a [new vignette describing how to download and curate data from GBIF](https://kmartinet.github.io/ssarp/articles/Get_Data.html).
- The documentation of `create_SAR` and `create_SpAR` now include additional information about how to extract the plot metadata for use in other plotting functions ([commit b8399e6](https://github.com/kmartinet/ssarp/commit/b8399e61273f81f9daaa86cacb78df602014fb94))

**Documentation response:**
- Thanks so much for reading the preprint of the `ssarp` manuscript! I really appreciate your comment that there is a lot of useful information there that should be added to the documentation.
- I added some more information about each speciation estimation method in the SpAR vignette.
- The documentation for `create_SpAR` now links to the SpAR vignette and mentions the three functions more explicitly.
- Technically all of the examples that use occurrence records in the package are using a user-provided dataset now that `SSARP::getData()' no longer exists! If these examples are not sufficient though, I am happy to add a section to the SAR vignette discussing custom datasets.
- I've added some clearer instructions for restricting occurrence records by geometry in the [new vignette describing gathering occurrence records from GBIF](https://kmartinet.github.io/ssarp/articles/Get_Data.html#restricting-occurrence-records-using-geometry)

**Input data response:**
- ```Patton_Anolis_Trimmed.tree``` is now in ```inst/extdata/``` ([In commit 7f29be1](https://github.com/kmartinet/SSARP/commit/7f29be1c9cdd2062118570c2c4485e4065ba271b))
- The code for accessing it across the package has been updated to ```read.tree(system.file("extdata", "Patton_Anolis_Trimmed.tree", package = "SSARP"))``` with three important exceptions: `test-estimate_BAMM()`, `test-estimate_DR()`, and `test-estimate_MS()`. The test-coverage.yaml workflow kept failing due to it not being able to find the file (even though it worked fine locally). I will continue to troubleshoot! My attempts are in these commits: [49cda1c](https://github.com/kmartinet/SSARP/commit/49cda1c6b111a8e4870098db3c0de58814255986), [2e8afb8](https://github.com/kmartinet/SSARP/commit/2e8afb8ad76167baaa2e4bc98883dacb637041b1), [8bc92c5](https://github.com/kmartinet/SSARP/commit/8bc92c5106c424b84f2999b37c9c959a19df5b1f), and [f79401e](https://github.com/kmartinet/SSARP/commit/f79401e8649b875d5140247c6dcca93a97f84cb2). If you have any guidance about how to ensure that the GitHub Actions workflows will not fail when using `extdata`, I would really appreciate your insight!
- Now that `SSARP::getData()` has been removed, the previous `rgbif::occ_search()` threshold is no longer an issue! I discuss methods for obtaining occurrence records using both `rgbif::occ_download()` (no limit) and `rgbif::occ_search()` in the [new vignette describing gathering data](https://kmartinet.github.io/ssarp/articles/Get_Data.html)

**Unit tests response:**
- New tests for calculations performed by ```ssarp::create_SAR()```, ```ssarp::creaate_SpAR()```, ```ssarp::estimate_BAMM()```, ```ssarp::estimate_DR()```, and ```ssarp::estimate_MS()``` have been written. ([commit d8d19fe](https://github.com/kmartinet/ssarp/commit/d8d19fedc5542d91a30844af0952c7281fc85d23)).
[…]
