### 2025-07-29 #685 @txm676
@jhollist @kmartinet

Thanks for your responses, they all look good to me and the new vignettes I think will be particularly useful. Only two further points:

* I'm not sure re your issue with extdata and ubuntu, sorry. The fact it works with windows and macs seems to imply there isn't actually an issue with how you've coded it. Did you update R or any other packages, as it could be something to do with an update to a package (or the source code) that is used the build and test processes (or the ubuntu packages). This has happened to me several times before.

* In response to this comment: I would prefer to only return the best-fit model for create_SAR and create_SpAR because I like the simplicity. Ultimately it is your package and I wouldn't want to push you into doing anything, but I would highlight that it can be quite frustrating as a user when you want to use something that you know is created by a function but isn't returned. My guess would be that the other models would be of use to some users (e.g., to generate comparative figures or model summary tables of their own), so it could be worth returning it all, even if just silently (i.e., not part of the printed output). But I leave it up to you!

Nice work on the package!
