### 2025-07-30 #685 @jhollist
First apologies for my delay in responding.  Been a bit chaotic and uncertain at work, but for the time being I'm still kicking...

@kmartinet thank you for the thorough responses.  They look solid to me.  @txm676 Thank you as well for the reply and thumbs up on the revisions.  For you both, I'll take a look at the extdata/Ubuntu thing to see if I have any insights.  Not promising anything!

@joelnitta could you take a look at @kmartinet's revisions and let us know your thoughts.

We are close!
