### 2025-07-30 #685 @jhollist
Quick update.  I do have an Ubuntu machine to test on and was able to recreated the error!  No solution yet, but at least I can test it...
