### 2025-08-01 084efe3
create_SAR and create_SpAR now return all models
