+++ R/create_SAR.R
@@ -15,9 +15,10 @@
 #' selection.  Default: 1
 #' @param visualize (boolean) Whether the plot should be displayed when the
 #' function is called. Default: FALSE
-#' @return A list of 4 including: the summary output, the regression
-#' object, the aggregated dataframe used to create the plot, and the AIC scores
-#' used in model selection
+#' @return A list of 5 including: the summary output, the regression
+#' object for the best-fit model, the aggregated dataframe used to create the 
+#' plot, the AIC scores used in model selection, 
+#' and all of the models created in model selection
 #' @examples
 #' # The GBIF key for the Anolis genus is 8782549
 #' # Read in example dataset filtered from:
@@ -79,6 +80,13 @@ create_SAR <- function(occurrences, npsi = 1, visualize = FALSE) {
   aic_scores[1] <- stats::AIC(linear)
   # Name the list element in case the user wants to examine the AIC scores
   names(aic_scores)[1] <- "Linear"
+  
+  # Empty list to populate with models
+  model_objs <- list()
+  # Linear model is already created above
+  model_objs[[1]] <- linear
+  # Name the list element
+  names(model_objs)[[1]] <- "Linear"
 
   # Create a segmented model for each level of npsi specified by the user
   # This only makes sense if the user does not specify that they want zero
@@ -95,6 +103,12 @@ create_SAR <- function(occurrences, npsi = 1, visualize = FALSE) {
       aic_scores[i + 1] <- stats::AIC(seg)
       # Name the list element
       names(aic_scores)[i + 1] <- paste0("BP", i)
+      
+      # Add to model list too
+      # Since the first index is always linear, use i+1
+      model_objs[[i + 1]] <- seg
+      # Name the list element
+      names(model_objs)[[i + 1]] <- paste0("BP", i)
     }
   }
 
@@ -123,7 +137,8 @@ create_SAR <- function(occurrences, npsi = 1, visualize = FALSE) {
       "summary" = summary_line,
       "linObj" = linear,
       "aggDF" = dat,
-      "AICscores" = aic_scores
+      "AICscores" = aic_scores,
+      "AllModels" = model_objs
     )
 
     class(result) <- "SAR"
@@ -164,7 +179,8 @@ create_SAR <- function(occurrences, npsi = 1, visualize = FALSE) {
       "summary" = summary_seg,
       "segObj" = seg,
       "aggDF" = dat,
-      "AICscores" = aic_scores
+      "AICscores" = aic_scores,
+      "AllModels" = model_objs
     )
 
     class(result) <- "SAR"
@@ -197,7 +213,8 @@ create_SAR <- function(occurrences, npsi = 1, visualize = FALSE) {
       "summary" = summary_seg,
       "segObj" = seg,
       "aggDF" = dat,
-      "AICscores" = aic_scores
+      "AICscores" = aic_scores,
+      "AllModels" = model_objs
     )
 
     class(result) <- "SAR"
