### 2025-08-01 cf210f5
Small wording change in plot.SAR documentation
