### 2025-08-01 #685 @kmartinet
Thanks so much, @txm676 ! That's a great point; I often get annoyed at returned objects not including all of the information I'm hoping to see. I've added the list of all candidate models to the returned object from `ssarp::create_SAR()` and `ssarp::create_SpAR()` ([commit 084efe3](https://github.com/kmartinet/ssarp/commit/084efe329c4fafa5abf8d578a6eb2c35f8056957)). That's a great point about updated packages being an issue for R-CMD-check too! My R version was out of date, so I just updated everything, made sure that `devtools::check()` passed locally, and made a small change to commit. R-CMD-check still failed on Ubuntu unfortunately (but continues to pass on Windows and MacOS).

And thanks so much, @jhollist , for working on the Ubuntu R-CMD-check issues! I appreciate your effort in running it on your personal machine and finding the open issue and PRs associated with the problems I'm having. I'm glad I'm not the only one!
