### 2025-08-08 ce7455a
Added get_pam_areas, for finding areas when a PAM is input
