### 2025-08-08 5415ace
Added find_pam_areas to the pkgdown site
