### 2025-09-16 #685 @kmartinet
Hi @jhollist ! I've officially changed all mixed-case functions (`create_SAR`, `create_SpAR`, `estimate_BAMM`, `estimate_DR`, and `estimate_MS`) to include only lowercase letters by default, with the original mixed-case spellings as aliases (commit [e9ae067](https://github.com/kmartinet/ssarp/commit/e9ae067de31c653c7ca988770007de17aef88dcf)). I also edited the vignettes and README to change the function references to the new lowercase versions.
