### 2025-09-17 #685 @jhollist
Everything looks great, @kmartinet! I think ssarp is now accepted!  I will get the process started.

There will be info in this issue on what needs to happen next once I do that.  If you have any questions or issues, just holler.

Thanks for your contribution and thank you to @joelnitta and @txm676 for your reviews and support of rOpenSci!
