### 2025-09-17 #685 @ropensci-review-bot
Approved! Thanks @kmartinet for submitting and @txm676, @joelnitta for your reviews! :grin:
